@@ -1,13 +1,14 @@
 name: fl_chart
 description: A highly customizable Flutter chart library that supports Line Chart, Bar Chart, Pie Chart, Scatter Chart, and Radar Chart.
-version: 0.63.0
+version: 0.65.0
 homepage: https://flchart.dev/
 repository: https://github.com/imaNNeo/fl_chart
 issue_tracker: https://github.com/imaNNeo/fl_chart/issues
 documentation: https://github.com/imaNNeo/fl_chart
 
 environment:
-  sdk: ^3.0.0
+  sdk: ^3.2.0
+  flutter: ^3.16.0
 
 funding:
   - https://github.com/sponsors/imaNNeo
@@ -19,12 +20,12 @@ dependencies:
     sdk: flutter
 
 dev_dependencies:
-  build_runner: ^2.4.2
+  build_runner: ^2.4.6
   flutter_test:
     sdk: flutter
-  mockito: ^5.4.0
+  mockito: ^5.4.2
   vector_math: ^2.1.4 #Added to use in some generated codes of mockito
-  very_good_analysis: ^5.0.0
+  very_good_analysis: ^5.1.0
 
 screenshots:
   - description: 'LineChartSample1 and LineChartSample2'
