@@ -1,6 +1,12 @@
+import 'package:dartx/dartx.dart';
+import 'package:fl_chart_app/cubits/app/app_cubit.dart';
 import 'package:fl_chart_app/presentation/resources/app_resources.dart';
+import 'package:fl_chart_app/urls.dart';
 import 'package:fl_chart_app/util/app_helper.dart';
+import 'package:fl_chart_app/util/app_utils.dart';
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 import 'fl_chart_banner.dart';
@@ -10,12 +16,14 @@ class AppMenu extends StatefulWidget {
   final List<ChartMenuItem> menuItems;
   final int currentSelectedIndex;
   final Function(int, ChartMenuItem) onItemSelected;
+  final VoidCallback? onBannerClicked;
 
   const AppMenu({
     Key? key,
     required this.menuItems,
     required this.currentSelectedIndex,
     required this.onItemSelected,
+    required this.onBannerClicked,
   }) : super(key: key);
 
   @override
@@ -25,16 +33,18 @@ class AppMenu extends StatefulWidget {
 class AppMenuState extends State<AppMenu> {
   @override
   Widget build(BuildContext context) {
-    const needToUpdateTheApp = 1 == 0;
     return Container(
       color: AppColors.itemsBackground,
       child: Column(
         children: [
-          const SafeArea(
+          SafeArea(
             child: AspectRatio(
               aspectRatio: 3,
               child: Center(
-                child: FlChartBanner(),
+                child: InkWell(
+                  onTap: widget.onBannerClicked,
+                  child: const FlChartBanner(),
+                ),
               ),
             ),
           ),
@@ -60,27 +70,79 @@ class AppMenuState extends State<AppMenu> {
               itemCount: widget.menuItems.length,
             ),
           ),
-          if (needToUpdateTheApp)
-            Container(
-              margin: const EdgeInsets.all(12),
-              child: Row(
-                children: [
-                  const Expanded(
-                    child: Padding(
-                      padding: EdgeInsets.all(10.0),
-                      child: Text(
-                        'FL Chart v 1.00 - update to get the latest features!',
-                        style: TextStyle(
-                          color: Colors.white,
-                          fontSize: 12,
+          const _AppVersionRow(),
+        ],
+      ),
+    );
+  }
+}
+
+class _AppVersionRow extends StatelessWidget {
+  const _AppVersionRow();
+
+  @override
+  Widget build(BuildContext context) {
+    return BlocBuilder<AppCubit, AppState>(builder: (context, state) {
+      if (state.appVersion.isNullOrBlank) {
+        return Container();
+      }
+      return Container(
+        margin: const EdgeInsets.all(12),
+        child: Row(
+          children: [
+            Expanded(
+              child: Padding(
+                padding: const EdgeInsets.all(10.0),
+                child: RichText(
+                  text: TextSpan(
+                    text: '',
+                    style: DefaultTextStyle.of(context).style,
+                    children: <TextSpan>[
+                      const TextSpan(text: 'App version: '),
+                      TextSpan(
+                        text: 'v${state.appVersion!}',
+                        style: const TextStyle(
+                          fontWeight: FontWeight.bold,
                         ),
                       ),
-                    ),
+                      if (state.usingFlChartVersion.isNotBlank) ...[
+                        TextSpan(
+                          text: '\nfl_chart: ',
+                          recognizer: TapGestureRecognizer()
+                            ..onTap = BlocProvider.of<AppCubit>(context)
+                                .onVersionClicked,
+                        ),
+                        TextSpan(
+                          text: 'v${state.usingFlChartVersion}',
+                          style: const TextStyle(
+                            fontWeight: FontWeight.bold,
+                          ),
+                          recognizer: TapGestureRecognizer()
+                            ..onTap = BlocProvider.of<AppCubit>(context)
+                                .onVersionClicked,
+                        ),
+                      ]
+                    ],
                   ),
-                  TextButton(
+                ),
+              ),
+            ),
+            state.availableVersionToUpdate.isNotBlank
+                ? TextButton(
                     onPressed: () {},
+                    child: Text(
+                      'Update to ${state.availableVersionToUpdate}',
+                      style: const TextStyle(
+                        color: AppColors.primary,
+                        fontSize: 12,
+                        fontWeight: FontWeight.bold,
+                      ),
+                    ),
+                  )
+                : TextButton(
+                    onPressed: () => AppUtils().tryToLaunchUrl(Urls.aboutUrl),
                     child: const Text(
-                      'Update',
+                      'About',
                       style: TextStyle(
                         color: AppColors.primary,
                         fontSize: 12,
@@ -88,12 +150,10 @@ class AppMenuState extends State<AppMenu> {
                       ),
                     ),
                   ),
-                ],
-              ),
-            )
-        ],
-      ),
-    );
+          ],
+        ),
+      );
+    });
   }
 }
 
