@@ -61,114 +61,431 @@ class ScatterChartSample1State extends State<ScatterChartSample1> {
   List<ScatterSpot> flutterLogoData() {
     return [
       /// section 1
-      ScatterSpot(20, 14.5, color: widget.blue1, radius: radius),
-      ScatterSpot(22, 16.5, color: widget.blue1, radius: radius),
-      ScatterSpot(24, 18.5, color: widget.blue1, radius: radius),
+      ScatterSpot(
+        20,
+        14.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue1, radius: radius),
+      ),
+      ScatterSpot(
+        20,
+        14.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue1, radius: radius),
+      ),
+      ScatterSpot(
+        22,
+        16.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue1, radius: radius),
+      ),
+      ScatterSpot(
+        24,
+        18.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue1, radius: radius),
+      ),
 
-      ScatterSpot(22, 12.5, color: widget.blue1, radius: radius),
-      ScatterSpot(24, 14.5, color: widget.blue1, radius: radius),
-      ScatterSpot(26, 16.5, color: widget.blue1, radius: radius),
+      ScatterSpot(
+        22,
+        12.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue1, radius: radius),
+      ),
+      ScatterSpot(
+        24,
+        14.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue1, radius: radius),
+      ),
+      ScatterSpot(
+        26,
+        16.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue1, radius: radius),
+      ),
 
-      ScatterSpot(24, 10.5, color: widget.blue1, radius: radius),
-      ScatterSpot(26, 12.5, color: widget.blue1, radius: radius),
-      ScatterSpot(28, 14.5, color: widget.blue1, radius: radius),
+      ScatterSpot(
+        24,
+        10.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue1, radius: radius),
+      ),
+      ScatterSpot(
+        26,
+        12.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue1, radius: radius),
+      ),
+      ScatterSpot(
+        28,
+        14.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue1, radius: radius),
+      ),
 
-      ScatterSpot(26, 8.5, color: widget.blue1, radius: radius),
-      ScatterSpot(28, 10.5, color: widget.blue1, radius: radius),
-      ScatterSpot(30, 12.5, color: widget.blue1, radius: radius),
+      ScatterSpot(
+        26,
+        8.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue1, radius: radius),
+      ),
+      ScatterSpot(
+        28,
+        10.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue1, radius: radius),
+      ),
+      ScatterSpot(
+        30,
+        12.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue1, radius: radius),
+      ),
 
-      ScatterSpot(28, 6.5, color: widget.blue1, radius: radius),
-      ScatterSpot(30, 8.5, color: widget.blue1, radius: radius),
-      ScatterSpot(32, 10.5, color: widget.blue1, radius: radius),
+      ScatterSpot(
+        28,
+        6.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue1, radius: radius),
+      ),
+      ScatterSpot(
+        30,
+        8.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue1, radius: radius),
+      ),
+      ScatterSpot(
+        32,
+        10.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue1, radius: radius),
+      ),
 
-      ScatterSpot(30, 4.5, color: widget.blue1, radius: radius),
-      ScatterSpot(32, 6.5, color: widget.blue1, radius: radius),
-      ScatterSpot(34, 8.5, color: widget.blue1, radius: radius),
+      ScatterSpot(
+        30,
+        4.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue1, radius: radius),
+      ),
+      ScatterSpot(
+        32,
+        6.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue1, radius: radius),
+      ),
+      ScatterSpot(
+        34,
+        8.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue1, radius: radius),
+      ),
 
-      ScatterSpot(34, 4.5, color: widget.blue1, radius: radius),
-      ScatterSpot(36, 6.5, color: widget.blue1, radius: radius),
+      ScatterSpot(
+        34,
+        4.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue1, radius: radius),
+      ),
+      ScatterSpot(
+        36,
+        6.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue1, radius: radius),
+      ),
 
-      ScatterSpot(38, 4.5, color: widget.blue1, radius: radius),
+      ScatterSpot(
+        38,
+        4.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue1, radius: radius),
+      ),
 
       /// section 2
-      ScatterSpot(20, 14.5, color: widget.blue2, radius: radius),
-      ScatterSpot(22, 12.5, color: widget.blue2, radius: radius),
-      ScatterSpot(24, 10.5, color: widget.blue2, radius: radius),
+      ScatterSpot(
+        20,
+        14.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        22,
+        12.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        24,
+        10.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
 
-      ScatterSpot(22, 16.5, color: widget.blue2, radius: radius),
-      ScatterSpot(24, 14.5, color: widget.blue2, radius: radius),
-      ScatterSpot(26, 12.5, color: widget.blue2, radius: radius),
+      ScatterSpot(
+        22,
+        16.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        24,
+        14.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        26,
+        12.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
 
-      ScatterSpot(24, 18.5, color: widget.blue2, radius: radius),
-      ScatterSpot(26, 16.5, color: widget.blue2, radius: radius),
-      ScatterSpot(28, 14.5, color: widget.blue2, radius: radius),
+      ScatterSpot(
+        24,
+        18.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        26,
+        16.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        28,
+        14.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
 
-      ScatterSpot(26, 20.5, color: widget.blue2, radius: radius),
-      ScatterSpot(28, 18.5, color: widget.blue2, radius: radius),
-      ScatterSpot(30, 16.5, color: widget.blue2, radius: radius),
+      ScatterSpot(
+        26,
+        20.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        28,
+        18.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        30,
+        16.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
 
-      ScatterSpot(28, 22.5, color: widget.blue2, radius: radius),
-      ScatterSpot(30, 20.5, color: widget.blue2, radius: radius),
-      ScatterSpot(32, 18.5, color: widget.blue2, radius: radius),
+      ScatterSpot(
+        28,
+        22.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        30,
+        20.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        32,
+        18.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
 
-      ScatterSpot(30, 24.5, color: widget.blue2, radius: radius),
-      ScatterSpot(32, 22.5, color: widget.blue2, radius: radius),
-      ScatterSpot(34, 20.5, color: widget.blue2, radius: radius),
+      ScatterSpot(
+        30,
+        24.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        32,
+        22.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        34,
+        20.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
 
-      ScatterSpot(34, 24.5, color: widget.blue2, radius: radius),
-      ScatterSpot(36, 22.5, color: widget.blue2, radius: radius),
+      ScatterSpot(
+        34,
+        24.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        36,
+        22.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
 
-      ScatterSpot(38, 24.5, color: widget.blue2, radius: radius),
+      ScatterSpot(
+        38,
+        24.5,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
 
       /// section 3
-      ScatterSpot(10, 25, color: widget.blue2, radius: radius),
-      ScatterSpot(12, 23, color: widget.blue2, radius: radius),
-      ScatterSpot(14, 21, color: widget.blue2, radius: radius),
+      ScatterSpot(
+        10,
+        25,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        12,
+        23,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        14,
+        21,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
 
-      ScatterSpot(12, 27, color: widget.blue2, radius: radius),
-      ScatterSpot(14, 25, color: widget.blue2, radius: radius),
-      ScatterSpot(16, 23, color: widget.blue2, radius: radius),
+      ScatterSpot(
+        12,
+        27,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        14,
+        25,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        16,
+        23,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
 
-      ScatterSpot(14, 29, color: widget.blue2, radius: radius),
-      ScatterSpot(16, 27, color: widget.blue2, radius: radius),
-      ScatterSpot(18, 25, color: widget.blue2, radius: radius),
+      ScatterSpot(
+        14,
+        29,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        16,
+        27,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        18,
+        25,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
 
-      ScatterSpot(16, 31, color: widget.blue2, radius: radius),
-      ScatterSpot(18, 29, color: widget.blue2, radius: radius),
-      ScatterSpot(20, 27, color: widget.blue2, radius: radius),
+      ScatterSpot(
+        16,
+        31,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        18,
+        29,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        20,
+        27,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
 
-      ScatterSpot(18, 33, color: widget.blue2, radius: radius),
-      ScatterSpot(20, 31, color: widget.blue2, radius: radius),
-      ScatterSpot(22, 29, color: widget.blue2, radius: radius),
+      ScatterSpot(
+        18,
+        33,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        20,
+        31,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        22,
+        29,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
 
-      ScatterSpot(20, 35, color: widget.blue2, radius: radius),
-      ScatterSpot(22, 33, color: widget.blue2, radius: radius),
-      ScatterSpot(24, 31, color: widget.blue2, radius: radius),
+      ScatterSpot(
+        20,
+        35,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        22,
+        33,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        24,
+        31,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
 
-      ScatterSpot(22, 37, color: widget.blue2, radius: radius),
-      ScatterSpot(24, 35, color: widget.blue2, radius: radius),
-      ScatterSpot(26, 33, color: widget.blue2, radius: radius),
+      ScatterSpot(
+        22,
+        37,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        24,
+        35,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        26,
+        33,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
 
-      ScatterSpot(24, 39, color: widget.blue2, radius: radius),
-      ScatterSpot(26, 37, color: widget.blue2, radius: radius),
-      ScatterSpot(28, 35, color: widget.blue2, radius: radius),
+      ScatterSpot(
+        24,
+        39,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        26,
+        37,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        28,
+        35,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
 
-      ScatterSpot(26, 41, color: widget.blue2, radius: radius),
-      ScatterSpot(28, 39, color: widget.blue2, radius: radius),
-      ScatterSpot(30, 37, color: widget.blue2, radius: radius),
+      ScatterSpot(
+        26,
+        41,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        28,
+        39,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        30,
+        37,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
 
-      ScatterSpot(28, 43, color: widget.blue2, radius: radius),
-      ScatterSpot(30, 41, color: widget.blue2, radius: radius),
-      ScatterSpot(32, 39, color: widget.blue2, radius: radius),
+      ScatterSpot(
+        28,
+        43,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        30,
+        41,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        32,
+        39,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
 
-      ScatterSpot(30, 45, color: widget.blue2, radius: radius),
-      ScatterSpot(32, 43, color: widget.blue2, radius: radius),
-      ScatterSpot(34, 41, color: widget.blue2, radius: radius),
+      ScatterSpot(
+        30,
+        45,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        32,
+        43,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        34,
+        41,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
 
-      ScatterSpot(34, 45, color: widget.blue2, radius: radius),
-      ScatterSpot(36, 43, color: widget.blue2, radius: radius),
+      ScatterSpot(
+        34,
+        45,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
+      ScatterSpot(
+        36,
+        43,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
 
-      ScatterSpot(38, 45, color: widget.blue2, radius: radius),
+      ScatterSpot(
+        38,
+        45,
+        dotPainter: FlDotCirclePainter(color: widget.blue2, radius: radius),
+      ),
     ];
   }
 
@@ -186,8 +503,10 @@ class ScatterChartSample1State extends State<ScatterChartSample1> {
       return ScatterSpot(
         (Random().nextDouble() * (maxX - 8)) + 4,
         (Random().nextDouble() * (maxY - 8)) + 4,
-        color: color,
-        radius: (Random().nextDouble() * 16) + 4,
+        dotPainter: FlDotCirclePainter(
+          color: color,
+          radius: (Random().nextDouble() * 16) + 4,
+        ),
       );
     });
   }
