@@ -796,205 +796,6 @@ class FlDotData with EquatableMixin {
       ];
 }
 
-/// This class contains the interface that all DotPainters should conform to.
-abstract class FlDotPainter with EquatableMixin {
-  const FlDotPainter();
-
-  /// This method should be overridden to draw the dot shape.
-  void draw(Canvas canvas, FlSpot spot, Offset offsetInCanvas);
-
-  /// This method should be overridden to return the size of the shape.
-  Size getSize(FlSpot spot);
-}
-
-/// This class is an implementation of a [FlDotPainter] that draws
-/// a circled shape
-class FlDotCirclePainter extends FlDotPainter {
-  /// The color of the circle is determined determined by [color],
-  /// [radius] determines the radius of the circle.
-  /// You can have a stroke line around the circle,
-  /// by setting the thickness with [strokeWidth],
-  /// and you can change the color of of the stroke with [strokeColor].
-  FlDotCirclePainter({
-    this.color = Colors.green,
-    double? radius,
-    this.strokeColor = const Color.fromRGBO(76, 175, 80, 1),
-    this.strokeWidth = 1.0,
-  }) : radius = radius ?? 4.0;
-
-  /// The fill color to use for the circle
-  Color color;
-
-  /// Customizes the radius of the circle
-  double radius;
-
-  /// The stroke color to use for the circle
-  Color strokeColor;
-
-  /// The stroke width to use for the circle
-  double strokeWidth;
-
-  /// Implementation of the parent class to draw the circle
-  @override
-  void draw(Canvas canvas, FlSpot spot, Offset offsetInCanvas) {
-    if (strokeWidth != 0.0 && strokeColor.opacity != 0.0) {
-      canvas.drawCircle(
-        offsetInCanvas,
-        radius + (strokeWidth / 2),
-        Paint()
-          ..color = strokeColor
-          ..strokeWidth = strokeWidth
-          ..style = PaintingStyle.stroke,
-      );
-    }
-    canvas.drawCircle(
-      offsetInCanvas,
-      radius,
-      Paint()
-        ..color = color
-        ..style = PaintingStyle.fill,
-    );
-  }
-
-  /// Implementation of the parent class to get the size of the circle
-  @override
-  Size getSize(FlSpot spot) {
-    return Size(radius * 2, radius * 2);
-  }
-
-  /// Used for equality check, see [EquatableMixin].
-  @override
-  List<Object?> get props => [
-        color,
-        radius,
-        strokeColor,
-        strokeWidth,
-      ];
-}
-
-/// This class is an implementation of a [FlDotPainter] that draws
-/// a squared shape
-class FlDotSquarePainter extends FlDotPainter {
-  /// The color of the square is determined determined by [color],
-  /// [size] determines the size of the square.
-  /// You can have a stroke line around the square,
-  /// by setting the thickness with [strokeWidth],
-  /// and you can change the color of of the stroke with [strokeColor].
-  FlDotSquarePainter({
-    this.color = Colors.green,
-    this.size = 4.0,
-    this.strokeColor = const Color.fromRGBO(76, 175, 80, 1),
-    this.strokeWidth = 1.0,
-  });
-
-  /// The fill color to use for the square
-  Color color;
-
-  /// Customizes the size of the square
-  double size;
-
-  /// The stroke color to use for the square
-  Color strokeColor;
-
-  /// The stroke width to use for the square
-  double strokeWidth;
-
-  /// Implementation of the parent class to draw the square
-  @override
-  void draw(Canvas canvas, FlSpot spot, Offset offsetInCanvas) {
-    if (strokeWidth != 0.0 && strokeColor.opacity != 0.0) {
-      canvas.drawRect(
-        Rect.fromCircle(
-          center: offsetInCanvas,
-          radius: (size / 2) + (strokeWidth / 2),
-        ),
-        Paint()
-          ..color = strokeColor
-          ..strokeWidth = strokeWidth
-          ..style = PaintingStyle.stroke,
-      );
-    }
-    canvas.drawRect(
-      Rect.fromCircle(
-        center: offsetInCanvas,
-        radius: size / 2,
-      ),
-      Paint()
-        ..color = color
-        ..style = PaintingStyle.fill,
-    );
-  }
-
-  /// Implementation of the parent class to get the size of the square
-  @override
-  Size getSize(FlSpot spot) {
-    return Size(size, size);
-  }
-
-  /// Used for equality check, see [EquatableMixin].
-  @override
-  List<Object?> get props => [
-        color,
-        size,
-        strokeColor,
-        strokeWidth,
-      ];
-}
-
-/// This class is an implementation of a [FlDotPainter] that draws
-/// a cross (X mark) shape
-class FlDotCrossPainter extends FlDotPainter {
-  /// The [color] and [width] properties determines the color and thickness of the cross shape,
-  /// [size] determines the width and height of the shape.
-  FlDotCrossPainter({
-    this.color = Colors.green,
-    this.size = 8.0,
-    this.width = 2.0,
-  });
-
-  /// The fill color to use for the X mark
-  Color color;
-
-  /// Determines size (width and height) of shape.
-  double size;
-
-  /// Determines thickness of X mark.
-  double width;
-
-  /// Implementation of the parent class to draw the cross
-  @override
-  void draw(Canvas canvas, FlSpot spot, Offset offsetInCanvas) {
-    final path = Path()
-      ..moveTo(offsetInCanvas.dx, offsetInCanvas.dy)
-      ..relativeMoveTo(-size / 2, -size / 2)
-      ..relativeLineTo(size, size)
-      ..moveTo(offsetInCanvas.dx, offsetInCanvas.dy)
-      ..relativeMoveTo(size / 2, -size / 2)
-      ..relativeLineTo(-size, size);
-
-    final paint = Paint()
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = width
-      ..color = color;
-
-    canvas.drawPath(path, paint);
-  }
-
-  /// Implementation of the parent class to get the size of the circle
-  @override
-  Size getSize(FlSpot spot) {
-    return Size(size, size);
-  }
-
-  /// Used for equality check, see [EquatableMixin].
-  @override
-  List<Object?> get props => [
-        color,
-        size,
-        width,
-      ];
-}
-
 /// It determines showing or hiding [FlDotData] on the spots.
 ///
 /// It gives you the checking [FlSpot] and you should decide to
@@ -1195,7 +996,7 @@ List<TouchedSpotIndicatorData> defaultTouchedIndicators(
       lineColor = _defaultGetDotColor(barData.spots[index], 0, barData);
     }
     const lineStrokeWidth = 4.0;
-    final flLine = FlLine(color: lineColor!, strokeWidth: lineStrokeWidth);
+    final flLine = FlLine(color: lineColor, strokeWidth: lineStrokeWidth);
 
     var dotSize = 10.0;
     if (barData.dotData.show) {
