@@ -411,16 +411,22 @@ class LineChartPainter extends AxisChartPainter<LineChartData> {
         }
       }
 
+      final indicatorLine = indicatorData.indicatorBelowLine;
       _touchLinePaint
-        ..color = indicatorData.indicatorBelowLine.color
-        ..strokeWidth = indicatorData.indicatorBelowLine.strokeWidth
+        ..setColorOrGradientForLine(
+          indicatorLine.color,
+          indicatorLine.gradient,
+          from: lineStart,
+          to: lineEnd,
+        )
+        ..strokeWidth = indicatorLine.strokeWidth
         ..transparentIfWidthIsZero();
 
       canvasWrapper.drawDashedLine(
         lineStart,
         lineEnd,
         _touchLinePaint,
-        indicatorData.indicatorBelowLine.dashArray,
+        indicatorLine.dashArray,
       );
 
       /// Draw the indicator dot
@@ -483,6 +489,9 @@ class LineChartPainter extends AxisChartPainter<LineChartData> {
     final y = getPixelY(barSpots[0].y, viewSize, holder);
     if (appendToPath == null) {
       path.moveTo(x, y);
+      if (size == 1) {
+        path.lineTo(x, y);
+      }
     } else {
       path.lineTo(x, y);
     }
@@ -752,17 +761,22 @@ class LineChartPainter extends AxisChartPainter<LineChartData> {
             );
           }
 
+          final lineStyle = barData.belowBarData.spotsLine.flLineStyle;
           _barAreaLinesPaint
-            ..color = barData.belowBarData.spotsLine.flLineStyle.color
-            ..strokeWidth =
-                barData.belowBarData.spotsLine.flLineStyle.strokeWidth
+            ..setColorOrGradientForLine(
+              lineStyle.color,
+              lineStyle.gradient,
+              from: from,
+              to: to,
+            )
+            ..strokeWidth = lineStyle.strokeWidth
             ..transparentIfWidthIsZero();
 
           canvasWrapper.drawDashedLine(
             from,
             to,
             _barAreaLinesPaint,
-            barData.belowBarData.spotsLine.flLineStyle.dashArray,
+            lineStyle.dashArray,
           );
         }
       }
@@ -841,17 +855,22 @@ class LineChartPainter extends AxisChartPainter<LineChartData> {
             );
           }
 
+          final lineStyle = barData.aboveBarData.spotsLine.flLineStyle;
           _barAreaLinesPaint
-            ..color = barData.aboveBarData.spotsLine.flLineStyle.color
-            ..strokeWidth =
-                barData.aboveBarData.spotsLine.flLineStyle.strokeWidth
+            ..setColorOrGradientForLine(
+              lineStyle.color,
+              lineStyle.gradient,
+              from: from,
+              to: to,
+            )
+            ..strokeWidth = lineStyle.strokeWidth
             ..transparentIfWidthIsZero();
 
           canvasWrapper.drawDashedLine(
             from,
             to,
             _barAreaLinesPaint,
-            barData.aboveBarData.spotsLine.flLineStyle.dashArray,
+            lineStyle.dashArray,
           );
         }
       }
@@ -891,6 +910,9 @@ class LineChartPainter extends AxisChartPainter<LineChartData> {
     LineChartBarData barData,
   ) {
     if (!barData.show || barData.shadow.color.opacity == 0.0) {
+      return;
+    }
+    if (barPath.computeMetrics().isEmpty) {
       return;
     }
 
@@ -993,7 +1015,7 @@ class LineChartPainter extends AxisChartPainter<LineChartData> {
         text: span,
         textAlign: tooltipItem.textAlign,
         textDirection: tooltipItem.textDirection,
-        textScaleFactor: holder.textScale,
+        textScaler: holder.textScaler,
       )..layout(maxWidth: tooltipData.maxContentWidth);
       drawingTextPainters.add(tp);
     }
@@ -1280,6 +1302,7 @@ class LineIndexDrawingInfo {
     this.spotIndex,
     this.indicatorData,
   );
+
   final LineChartBarData line;
   final int lineIndex;
   final FlSpot spot;
