@@ -26,7 +26,7 @@ class PieChartLeaf extends MultiChildRenderObjectWidget {
         context,
         data,
         targetData,
-        MediaQuery.of(context).textScaleFactor,
+        MediaQuery.of(context).textScaler,
       );
 
   @override
@@ -34,7 +34,7 @@ class PieChartLeaf extends MultiChildRenderObjectWidget {
     renderObject
       ..data = data
       ..targetData = targetData
-      ..textScale = MediaQuery.of(context).textScaleFactor
+      ..textScaler = MediaQuery.of(context).textScaler
       ..buildContext = context;
   }
 }
@@ -50,10 +50,10 @@ class RenderPieChart extends RenderBaseChart<PieTouchResponse>
     BuildContext context,
     PieChartData data,
     PieChartData targetData,
-    double textScale,
+    TextScaler textScaler,
   )   : _data = data,
         _targetData = targetData,
-        _textScale = textScale,
+        _textScaler = textScaler,
         super(targetData.pieTouchData, context);
 
   PieChartData get data => _data;
@@ -77,12 +77,12 @@ class RenderPieChart extends RenderBaseChart<PieTouchResponse>
     markNeedsLayout();
   }
 
-  double get textScale => _textScale;
-  double _textScale;
+  TextScaler get textScaler => _textScaler;
+  TextScaler _textScaler;
 
-  set textScale(double value) {
-    if (_textScale == value) return;
-    _textScale = value;
+  set textScaler(TextScaler value) {
+    if (_textScaler == value) return;
+    _textScaler = value;
     markNeedsPaint();
   }
 
@@ -93,9 +93,8 @@ class RenderPieChart extends RenderBaseChart<PieTouchResponse>
   @visibleForTesting
   PieChartPainter painter = PieChartPainter();
 
-  PaintHolder<PieChartData> get paintHolder {
-    return PaintHolder(data, targetData, textScale);
-  }
+  PaintHolder<PieChartData> get paintHolder =>
+      PaintHolder(data, targetData, textScaler);
 
   @override
   void setupParentData(RenderBox child) {
