@@ -4,6 +4,22 @@ extension BarChartDataExtension on BarChartData {
   List<double> calculateGroupsX(double viewWidth) {
     assert(barGroups.isNotEmpty);
     final groupsX = List<double>.filled(barGroups.length, 0);
+
+    var sumWidth =
+        barGroups.map((group) => group.width).reduce((a, b) => a + b);
+    final spaceAvailable = viewWidth - sumWidth;
+
+    void spaceEvenly() {
+      final eachSpace = spaceAvailable / (barGroups.length + 1);
+      var tempX = 0.0;
+      barGroups.asMap().forEach((i, group) {
+        tempX += eachSpace;
+        tempX += group.width / 2;
+        groupsX[i] = tempX;
+        tempX += group.width / 2;
+      });
+    }
+
     switch (alignment) {
       case BarChartAlignment.start:
         var tempX = 0.0;
@@ -14,10 +30,11 @@ extension BarChartDataExtension on BarChartData {
           final groupSpace = i == barGroups.length - 1 ? 0 : groupsSpace;
           tempX += group.width + groupSpace;
         }
+        if (tempX > viewWidth) {
+          spaceEvenly();
+        }
 
       case BarChartAlignment.end:
-        var sumWidth =
-            barGroups.map((group) => group.width).reduce((a, b) => a + b);
         sumWidth += groupsSpace * (barGroups.length - 1);
         final horizontalMargin = viewWidth - sumWidth;
 
@@ -29,10 +46,11 @@ extension BarChartDataExtension on BarChartData {
           final groupSpace = i == barGroups.length - 1 ? 0 : groupsSpace;
           tempX += group.width + groupSpace;
         }
+        if (tempX > viewWidth) {
+          spaceEvenly();
+        }
 
       case BarChartAlignment.center:
-        var sumWidth =
-            barGroups.map((group) => group.width).reduce((a, b) => a + b);
         sumWidth += groupsSpace * (barGroups.length - 1);
         final horizontalMargin = (viewWidth - sumWidth) / 2;
 
@@ -44,11 +62,11 @@ extension BarChartDataExtension on BarChartData {
           final groupSpace = i == barGroups.length - 1 ? 0 : groupsSpace;
           tempX += group.width + groupSpace;
         }
+        if (tempX > viewWidth) {
+          spaceEvenly();
+        }
 
       case BarChartAlignment.spaceBetween:
-        final sumWidth =
-            barGroups.map((group) => group.width).reduce((a, b) => a + b);
-        final spaceAvailable = viewWidth - sumWidth;
         final eachSpace = spaceAvailable / (barGroups.length - 1);
 
         var tempX = 0.0;
@@ -62,9 +80,6 @@ extension BarChartDataExtension on BarChartData {
         });
 
       case BarChartAlignment.spaceAround:
-        final sumWidth =
-            barGroups.map((group) => group.width).reduce((a, b) => a + b);
-        final spaceAvailable = viewWidth - sumWidth;
         final eachSpace = spaceAvailable / (barGroups.length * 2);
 
         var tempX = 0.0;
@@ -77,18 +92,7 @@ extension BarChartDataExtension on BarChartData {
         });
 
       case BarChartAlignment.spaceEvenly:
-        final sumWidth =
-            barGroups.map((group) => group.width).reduce((a, b) => a + b);
-        final spaceAvailable = viewWidth - sumWidth;
-        final eachSpace = spaceAvailable / (barGroups.length + 1);
-
-        var tempX = 0.0;
-        barGroups.asMap().forEach((i, group) {
-          tempX += eachSpace;
-          tempX += group.width / 2;
-          groupsX[i] = tempX;
-          tempX += group.width / 2;
-        });
+        spaceEvenly();
     }
 
     return groupsX;
