@@ -1,4 +1,4 @@
-// Mocks generated by Mockito 5.4.0 from annotations
+// Mocks generated by Mockito 5.4.3 from annotations
 // in fl_chart/test/chart/bar_chart/bar_chart_painter_test.dart.
 // Do not manually edit this file.
 
@@ -12,11 +12,14 @@ import 'package:fl_chart/src/utils/utils.dart' as _i8;
 import 'package:flutter/cupertino.dart' as _i3;
 import 'package:flutter/foundation.dart' as _i4;
 import 'package:mockito/mockito.dart' as _i1;
+import 'package:mockito/src/dummies.dart' as _i9;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
 // ignore_for_file: avoid_setters_without_getters
 // ignore_for_file: comment_references
+// ignore_for_file: deprecated_member_use
+// ignore_for_file: deprecated_member_use_from_same_package
 // ignore_for_file: implementation_imports
 // ignore_for_file: invalid_use_of_visible_for_testing_member
 // ignore_for_file: prefer_const_constructors
@@ -155,6 +158,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void saveLayer(
     _i2.Rect? bounds,
@@ -170,6 +174,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void restore() => super.noSuchMethod(
         Invocation.method(
@@ -178,6 +183,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void restoreToCount(int? count) => super.noSuchMethod(
         Invocation.method(
@@ -186,6 +192,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   int getSaveCount() => (super.noSuchMethod(
         Invocation.method(
@@ -194,6 +201,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValue: 0,
       ) as int);
+
   @override
   void translate(
     double? dx,
@@ -209,6 +217,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void scale(
     double? sx, [
@@ -224,6 +233,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void rotate(double? radians) => super.noSuchMethod(
         Invocation.method(
@@ -232,6 +242,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void skew(
     double? sx,
@@ -247,6 +258,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void transform(_i5.Float64List? matrix4) => super.noSuchMethod(
         Invocation.method(
@@ -255,6 +267,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i5.Float64List getTransform() => (super.noSuchMethod(
         Invocation.method(
@@ -263,6 +276,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValue: _i5.Float64List(0),
       ) as _i5.Float64List);
+
   @override
   void clipRect(
     _i2.Rect? rect, {
@@ -280,6 +294,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipRRect(
     _i2.RRect? rrect, {
@@ -293,6 +308,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipPath(
     _i2.Path? path, {
@@ -306,6 +322,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i2.Rect getLocalClipBounds() => (super.noSuchMethod(
         Invocation.method(
@@ -320,6 +337,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
           ),
         ),
       ) as _i2.Rect);
+
   @override
   _i2.Rect getDestinationClipBounds() => (super.noSuchMethod(
         Invocation.method(
@@ -334,6 +352,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
           ),
         ),
       ) as _i2.Rect);
+
   @override
   void drawColor(
     _i2.Color? color,
@@ -349,6 +368,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawLine(
     _i2.Offset? p1,
@@ -366,6 +386,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPaint(_i2.Paint? paint) => super.noSuchMethod(
         Invocation.method(
@@ -374,6 +395,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRect(
     _i2.Rect? rect,
@@ -389,6 +411,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRRect(
     _i2.RRect? rrect,
@@ -404,6 +427,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawDRRect(
     _i2.RRect? outer,
@@ -421,6 +445,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawOval(
     _i2.Rect? rect,
@@ -436,6 +461,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawCircle(
     _i2.Offset? c,
@@ -453,6 +479,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawArc(
     _i2.Rect? rect,
@@ -474,6 +501,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPath(
     _i2.Path? path,
@@ -489,6 +517,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawImage(
     _i2.Image? image,
@@ -506,6 +535,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawImageRect(
     _i2.Image? image,
@@ -525,6 +555,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawImageNine(
     _i2.Image? image,
@@ -544,6 +575,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPicture(_i2.Picture? picture) => super.noSuchMethod(
         Invocation.method(
@@ -552,6 +584,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawParagraph(
     _i2.Paragraph? paragraph,
@@ -567,6 +600,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPoints(
     _i2.PointMode? pointMode,
@@ -584,6 +618,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRawPoints(
     _i2.PointMode? pointMode,
@@ -601,6 +636,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawVertices(
     _i2.Vertices? vertices,
@@ -618,6 +654,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawAtlas(
     _i2.Image? atlas,
@@ -643,6 +680,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRawAtlas(
     _i2.Image? atlas,
@@ -668,6 +706,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawShadow(
     _i2.Path? path,
@@ -705,6 +744,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
           Invocation.getter(#canvas),
         ),
       ) as _i2.Canvas);
+
   @override
   _i2.Size get size => (super.noSuchMethod(
         Invocation.getter(#size),
@@ -713,6 +753,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
           Invocation.getter(#size),
         ),
       ) as _i2.Size);
+
   @override
   void drawRRect(
     _i2.RRect? rrect,
@@ -728,6 +769,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void save() => super.noSuchMethod(
         Invocation.method(
@@ -736,6 +778,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void restore() => super.noSuchMethod(
         Invocation.method(
@@ -744,6 +787,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipRect(
     _i2.Rect? rect, {
@@ -761,6 +805,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void translate(
     double? dx,
@@ -776,6 +821,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void rotate(double? radius) => super.noSuchMethod(
         Invocation.method(
@@ -784,6 +830,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPath(
     _i2.Path? path,
@@ -799,6 +846,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void saveLayer(
     _i2.Rect? bounds,
@@ -814,6 +862,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPicture(_i2.Picture? picture) => super.noSuchMethod(
         Invocation.method(
@@ -822,6 +871,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawImage(
     _i2.Image? image,
@@ -839,6 +889,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipPath(
     _i2.Path? path, {
@@ -852,6 +903,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRect(
     _i2.Rect? rect,
@@ -867,6 +919,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawLine(
     _i2.Offset? p1,
@@ -884,6 +937,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawCircle(
     _i2.Offset? center,
@@ -901,6 +955,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawArc(
     _i2.Rect? rect,
@@ -922,6 +977,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawText(
     _i3.TextPainter? tp,
@@ -939,6 +995,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawDot(
     _i7.FlDotPainter? painter,
@@ -956,6 +1013,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRotated({
     required _i2.Size? size,
@@ -978,6 +1036,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawDashedLine(
     _i2.Offset? from,
@@ -1015,16 +1074,19 @@ class MockBuildContext extends _i1.Mock implements _i3.BuildContext {
           Invocation.getter(#widget),
         ),
       ) as _i3.Widget);
+
   @override
   bool get mounted => (super.noSuchMethod(
         Invocation.getter(#mounted),
         returnValue: false,
       ) as bool);
+
   @override
   bool get debugDoingBuild => (super.noSuchMethod(
         Invocation.getter(#debugDoingBuild),
         returnValue: false,
       ) as bool);
+
   @override
   _i3.InheritedWidget dependOnInheritedElement(
     _i3.InheritedElement? ancestor, {
@@ -1045,6 +1107,7 @@ class MockBuildContext extends _i1.Mock implements _i3.BuildContext {
           ),
         ),
       ) as _i3.InheritedWidget);
+
   @override
   void visitAncestorElements(_i3.ConditionalElementVisitor? visitor) =>
       super.noSuchMethod(
@@ -1054,6 +1117,7 @@ class MockBuildContext extends _i1.Mock implements _i3.BuildContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void visitChildElements(_i3.ElementVisitor? visitor) => super.noSuchMethod(
         Invocation.method(
@@ -1062,6 +1126,7 @@ class MockBuildContext extends _i1.Mock implements _i3.BuildContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void dispatchNotification(_i3.Notification? notification) =>
       super.noSuchMethod(
@@ -1071,6 +1136,7 @@ class MockBuildContext extends _i1.Mock implements _i3.BuildContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i3.DiagnosticsNode describeElement(
     String? name, {
@@ -1091,6 +1157,7 @@ class MockBuildContext extends _i1.Mock implements _i3.BuildContext {
           ),
         ),
       ) as _i3.DiagnosticsNode);
+
   @override
   _i3.DiagnosticsNode describeWidget(
     String? name, {
@@ -1111,6 +1178,7 @@ class MockBuildContext extends _i1.Mock implements _i3.BuildContext {
           ),
         ),
       ) as _i3.DiagnosticsNode);
+
   @override
   List<_i3.DiagnosticsNode> describeMissingAncestor(
           {required Type? expectedAncestorType}) =>
@@ -1122,6 +1190,7 @@ class MockBuildContext extends _i1.Mock implements _i3.BuildContext {
         ),
         returnValue: <_i3.DiagnosticsNode>[],
       ) as List<_i3.DiagnosticsNode>);
+
   @override
   _i3.DiagnosticsNode describeOwnershipChain(String? name) =>
       (super.noSuchMethod(
@@ -1155,6 +1224,7 @@ class MockUtils extends _i1.Mock implements _i8.Utils {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   double degrees(double? radians) => (super.noSuchMethod(
         Invocation.method(
@@ -1163,6 +1233,7 @@ class MockUtils extends _i1.Mock implements _i8.Utils {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   _i2.Size getDefaultSize(_i2.Size? screenSize) => (super.noSuchMethod(
         Invocation.method(
@@ -1177,6 +1248,7 @@ class MockUtils extends _i1.Mock implements _i8.Utils {
           ),
         ),
       ) as _i2.Size);
+
   @override
   double translateRotatedPosition(
     double? size,
@@ -1192,6 +1264,7 @@ class MockUtils extends _i1.Mock implements _i8.Utils {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   _i2.Offset calculateRotationOffset(
     _i2.Size? size,
@@ -1216,6 +1289,7 @@ class MockUtils extends _i1.Mock implements _i8.Utils {
           ),
         ),
       ) as _i2.Offset);
+
   @override
   _i3.BorderRadius? normalizeBorderRadius(
     _i3.BorderRadius? borderRadius,
@@ -1228,6 +1302,7 @@ class MockUtils extends _i1.Mock implements _i8.Utils {
           width,
         ],
       )) as _i3.BorderRadius?);
+
   @override
   _i3.BorderSide normalizeBorderSide(
     _i3.BorderSide? borderSide,
@@ -1252,6 +1327,7 @@ class MockUtils extends _i1.Mock implements _i8.Utils {
           ),
         ),
       ) as _i3.BorderSide);
+
   @override
   double getEfficientInterval(
     double? axisViewSize,
@@ -1269,6 +1345,7 @@ class MockUtils extends _i1.Mock implements _i8.Utils {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   double roundInterval(double? input) => (super.noSuchMethod(
         Invocation.method(
@@ -1277,14 +1354,44 @@ class MockUtils extends _i1.Mock implements _i8.Utils {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
-  String formatNumber(double? number) => (super.noSuchMethod(
+  int getFractionDigits(double? value) => (super.noSuchMethod(
+        Invocation.method(
+          #getFractionDigits,
+          [value],
+        ),
+        returnValue: 0,
+      ) as int);
+
+  @override
+  String formatNumber(
+    double? axisMin,
+    double? axisMax,
+    double? axisValue,
+  ) =>
+      (super.noSuchMethod(
         Invocation.method(
           #formatNumber,
-          [number],
+          [
+            axisMin,
+            axisMax,
+            axisValue,
+          ],
         ),
-        returnValue: '',
+        returnValue: _i9.dummyValue<String>(
+          this,
+          Invocation.method(
+            #formatNumber,
+            [
+              axisMin,
+              axisMax,
+              axisValue,
+            ],
+          ),
+        ),
       ) as String);
+
   @override
   _i3.TextStyle getThemeAwareTextStyle(
     _i3.BuildContext? context,
@@ -1309,6 +1416,7 @@ class MockUtils extends _i1.Mock implements _i8.Utils {
           ),
         ),
       ) as _i3.TextStyle);
+
   @override
   double getBestInitialIntervalValue(
     double? min,
@@ -1328,6 +1436,7 @@ class MockUtils extends _i1.Mock implements _i8.Utils {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   double convertRadiusToSigma(double? radius) => (super.noSuchMethod(
         Invocation.method(
