@@ -1,4 +1,4 @@
-// Mocks generated by Mockito 5.4.0 from annotations
+// Mocks generated by Mockito 5.4.3 from annotations
 // in fl_chart/test/chart/bar_chart/bar_chart_renderer_test.dart.
 // Do not manually edit this file.
 
@@ -23,6 +23,8 @@ import 'package:mockito/mockito.dart' as _i1;
 // ignore_for_file: avoid_redundant_argument_values
 // ignore_for_file: avoid_setters_without_getters
 // ignore_for_file: comment_references
+// ignore_for_file: deprecated_member_use
+// ignore_for_file: deprecated_member_use_from_same_package
 // ignore_for_file: implementation_imports
 // ignore_for_file: invalid_use_of_visible_for_testing_member
 // ignore_for_file: prefer_const_constructors
@@ -153,6 +155,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void saveLayer(
     _i2.Rect? bounds,
@@ -168,6 +171,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void restore() => super.noSuchMethod(
         Invocation.method(
@@ -176,6 +180,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void restoreToCount(int? count) => super.noSuchMethod(
         Invocation.method(
@@ -184,6 +189,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   int getSaveCount() => (super.noSuchMethod(
         Invocation.method(
@@ -192,6 +198,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValue: 0,
       ) as int);
+
   @override
   void translate(
     double? dx,
@@ -207,6 +214,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void scale(
     double? sx, [
@@ -222,6 +230,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void rotate(double? radians) => super.noSuchMethod(
         Invocation.method(
@@ -230,6 +239,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void skew(
     double? sx,
@@ -245,6 +255,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void transform(_i7.Float64List? matrix4) => super.noSuchMethod(
         Invocation.method(
@@ -253,6 +264,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i7.Float64List getTransform() => (super.noSuchMethod(
         Invocation.method(
@@ -261,6 +273,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValue: _i7.Float64List(0),
       ) as _i7.Float64List);
+
   @override
   void clipRect(
     _i2.Rect? rect, {
@@ -278,6 +291,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipRRect(
     _i2.RRect? rrect, {
@@ -291,6 +305,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipPath(
     _i2.Path? path, {
@@ -304,6 +319,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i2.Rect getLocalClipBounds() => (super.noSuchMethod(
         Invocation.method(
@@ -318,6 +334,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
           ),
         ),
       ) as _i2.Rect);
+
   @override
   _i2.Rect getDestinationClipBounds() => (super.noSuchMethod(
         Invocation.method(
@@ -332,6 +349,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
           ),
         ),
       ) as _i2.Rect);
+
   @override
   void drawColor(
     _i2.Color? color,
@@ -347,6 +365,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawLine(
     _i2.Offset? p1,
@@ -364,6 +383,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPaint(_i2.Paint? paint) => super.noSuchMethod(
         Invocation.method(
@@ -372,6 +392,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRect(
     _i2.Rect? rect,
@@ -387,6 +408,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRRect(
     _i2.RRect? rrect,
@@ -402,6 +424,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawDRRect(
     _i2.RRect? outer,
@@ -419,6 +442,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawOval(
     _i2.Rect? rect,
@@ -434,6 +458,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawCircle(
     _i2.Offset? c,
@@ -451,6 +476,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawArc(
     _i2.Rect? rect,
@@ -472,6 +498,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPath(
     _i2.Path? path,
@@ -487,6 +514,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawImage(
     _i2.Image? image,
@@ -504,6 +532,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawImageRect(
     _i2.Image? image,
@@ -523,6 +552,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawImageNine(
     _i2.Image? image,
@@ -542,6 +572,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPicture(_i2.Picture? picture) => super.noSuchMethod(
         Invocation.method(
@@ -550,6 +581,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawParagraph(
     _i2.Paragraph? paragraph,
@@ -565,6 +597,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPoints(
     _i2.PointMode? pointMode,
@@ -582,6 +615,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRawPoints(
     _i2.PointMode? pointMode,
@@ -599,6 +633,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawVertices(
     _i2.Vertices? vertices,
@@ -616,6 +651,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawAtlas(
     _i2.Image? atlas,
@@ -641,6 +677,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRawAtlas(
     _i2.Image? atlas,
@@ -666,6 +703,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawShadow(
     _i2.Path? path,
@@ -703,6 +741,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           Invocation.getter(#estimatedBounds),
         ),
       ) as _i2.Rect);
+
   @override
   _i2.Canvas get canvas => (super.noSuchMethod(
         Invocation.getter(#canvas),
@@ -711,6 +750,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           Invocation.getter(#canvas),
         ),
       ) as _i2.Canvas);
+
   @override
   void paintChild(
     _i3.RenderObject? child,
@@ -726,6 +766,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void appendLayer(_i4.Layer? layer) => super.noSuchMethod(
         Invocation.method(
@@ -734,6 +775,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i2.VoidCallback addCompositionCallback(_i4.CompositionCallback? callback) =>
       (super.noSuchMethod(
@@ -743,6 +785,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValue: () {},
       ) as _i2.VoidCallback);
+
   @override
   void stopRecordingIfNeeded() => super.noSuchMethod(
         Invocation.method(
@@ -751,6 +794,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void setIsComplexHint() => super.noSuchMethod(
         Invocation.method(
@@ -759,6 +803,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void setWillChangeHint() => super.noSuchMethod(
         Invocation.method(
@@ -767,6 +812,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void addLayer(_i4.Layer? layer) => super.noSuchMethod(
         Invocation.method(
@@ -775,6 +821,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void pushLayer(
     _i4.ContainerLayer? childLayer,
@@ -794,6 +841,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i3.PaintingContext createChildContext(
     _i4.ContainerLayer? childLayer,
@@ -818,6 +866,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           ),
         ),
       ) as _i3.PaintingContext);
+
   @override
   _i4.ClipRectLayer? pushClipRect(
     bool? needsCompositing,
@@ -840,6 +889,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           #oldLayer: oldLayer,
         },
       )) as _i4.ClipRectLayer?);
+
   @override
   _i4.ClipRRectLayer? pushClipRRect(
     bool? needsCompositing,
@@ -864,6 +914,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           #oldLayer: oldLayer,
         },
       )) as _i4.ClipRRectLayer?);
+
   @override
   _i4.ClipPathLayer? pushClipPath(
     bool? needsCompositing,
@@ -888,6 +939,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           #oldLayer: oldLayer,
         },
       )) as _i4.ClipPathLayer?);
+
   @override
   _i4.ColorFilterLayer pushColorFilter(
     _i2.Offset? offset,
@@ -918,6 +970,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           ),
         ),
       ) as _i4.ColorFilterLayer);
+
   @override
   _i4.TransformLayer? pushTransform(
     bool? needsCompositing,
@@ -936,6 +989,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ],
         {#oldLayer: oldLayer},
       )) as _i4.TransformLayer?);
+
   @override
   _i4.OpacityLayer pushOpacity(
     _i2.Offset? offset,
@@ -966,6 +1020,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           ),
         ),
       ) as _i4.OpacityLayer);
+
   @override
   void clipPathAndPaint(
     _i2.Path? path,
@@ -985,6 +1040,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipRRectAndPaint(
     _i2.RRect? rrect,
@@ -1004,6 +1060,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipRectAndPaint(
     _i2.Rect? rect,
@@ -1041,16 +1098,19 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
           Invocation.getter(#widget),
         ),
       ) as _i6.Widget);
+
   @override
   bool get mounted => (super.noSuchMethod(
         Invocation.getter(#mounted),
         returnValue: false,
       ) as bool);
+
   @override
   bool get debugDoingBuild => (super.noSuchMethod(
         Invocation.getter(#debugDoingBuild),
         returnValue: false,
       ) as bool);
+
   @override
   _i6.InheritedWidget dependOnInheritedElement(
     _i6.InheritedElement? ancestor, {
@@ -1071,6 +1131,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
           ),
         ),
       ) as _i6.InheritedWidget);
+
   @override
   void visitAncestorElements(_i6.ConditionalElementVisitor? visitor) =>
       super.noSuchMethod(
@@ -1080,6 +1141,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void visitChildElements(_i6.ElementVisitor? visitor) => super.noSuchMethod(
         Invocation.method(
@@ -1088,6 +1150,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void dispatchNotification(_i9.Notification? notification) =>
       super.noSuchMethod(
@@ -1097,6 +1160,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i5.DiagnosticsNode describeElement(
     String? name, {
@@ -1117,6 +1181,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
           ),
         ),
       ) as _i5.DiagnosticsNode);
+
   @override
   _i5.DiagnosticsNode describeWidget(
     String? name, {
@@ -1137,6 +1202,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
           ),
         ),
       ) as _i5.DiagnosticsNode);
+
   @override
   List<_i5.DiagnosticsNode> describeMissingAncestor(
           {required Type? expectedAncestorType}) =>
@@ -1148,6 +1214,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
         ),
         returnValue: <_i5.DiagnosticsNode>[],
       ) as List<_i5.DiagnosticsNode>);
+
   @override
   _i5.DiagnosticsNode describeOwnershipChain(String? name) =>
       (super.noSuchMethod(
@@ -1190,6 +1257,7 @@ class MockBarChartPainter extends _i1.Mock implements _i10.BarChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   List<_i10.GroupBarsPosition> calculateGroupAndBarsPosition(
     _i2.Size? viewSize,
@@ -1207,6 +1275,7 @@ class MockBarChartPainter extends _i1.Mock implements _i10.BarChartPainter {
         ),
         returnValue: <_i10.GroupBarsPosition>[],
       ) as List<_i10.GroupBarsPosition>);
+
   @override
   void drawBars(
     _i11.CanvasWrapper? canvasWrapper,
@@ -1224,6 +1293,7 @@ class MockBarChartPainter extends _i1.Mock implements _i10.BarChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawTouchTooltip(
     _i6.BuildContext? context,
@@ -1253,6 +1323,7 @@ class MockBarChartPainter extends _i1.Mock implements _i10.BarChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawStackItemBorderStroke(
     _i11.CanvasWrapper? canvasWrapper,
@@ -1280,6 +1351,7 @@ class MockBarChartPainter extends _i1.Mock implements _i10.BarChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i13.BarTouchedSpot? handleTouch(
     _i2.Offset? localPosition,
@@ -1294,6 +1366,7 @@ class MockBarChartPainter extends _i1.Mock implements _i10.BarChartPainter {
           holder,
         ],
       )) as _i13.BarTouchedSpot?);
+
   @override
   void drawGrid(
     _i11.CanvasWrapper? canvasWrapper,
@@ -1309,6 +1382,7 @@ class MockBarChartPainter extends _i1.Mock implements _i10.BarChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawBackground(
     _i11.CanvasWrapper? canvasWrapper,
@@ -1324,6 +1398,7 @@ class MockBarChartPainter extends _i1.Mock implements _i10.BarChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRangeAnnotation(
     _i11.CanvasWrapper? canvasWrapper,
@@ -1339,6 +1414,7 @@ class MockBarChartPainter extends _i1.Mock implements _i10.BarChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawExtraLines(
     _i6.BuildContext? context,
@@ -1356,6 +1432,7 @@ class MockBarChartPainter extends _i1.Mock implements _i10.BarChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawHorizontalLines(
     _i6.BuildContext? context,
@@ -1375,6 +1452,7 @@ class MockBarChartPainter extends _i1.Mock implements _i10.BarChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawVerticalLines(
     _i6.BuildContext? context,
@@ -1394,6 +1472,7 @@ class MockBarChartPainter extends _i1.Mock implements _i10.BarChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   double getPixelX(
     double? spotX,
@@ -1411,6 +1490,7 @@ class MockBarChartPainter extends _i1.Mock implements _i10.BarChartPainter {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   double getPixelY(
     double? spotY,
@@ -1428,6 +1508,7 @@ class MockBarChartPainter extends _i1.Mock implements _i10.BarChartPainter {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   double getTooltipLeft(
     double? dx,
