@@ -1,4 +1,4 @@
-// Mocks generated by Mockito 5.4.0 from annotations
+// Mocks generated by Mockito 5.4.3 from annotations
 // in fl_chart/test/chart/line_chart/line_chart_painter_test.dart.
 // Do not manually edit this file.
 
@@ -8,18 +8,21 @@ import 'dart:ui' as _i2;
 
 import 'package:fl_chart/fl_chart.dart' as _i7;
 import 'package:fl_chart/src/chart/base/base_chart/base_chart_painter.dart'
-    as _i10;
-import 'package:fl_chart/src/chart/line_chart/line_chart_painter.dart' as _i9;
+    as _i11;
+import 'package:fl_chart/src/chart/line_chart/line_chart_painter.dart' as _i10;
 import 'package:fl_chart/src/utils/canvas_wrapper.dart' as _i6;
 import 'package:fl_chart/src/utils/utils.dart' as _i8;
 import 'package:flutter/cupertino.dart' as _i3;
 import 'package:flutter/foundation.dart' as _i4;
 import 'package:mockito/mockito.dart' as _i1;
+import 'package:mockito/src/dummies.dart' as _i9;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
 // ignore_for_file: avoid_setters_without_getters
 // ignore_for_file: comment_references
+// ignore_for_file: deprecated_member_use
+// ignore_for_file: deprecated_member_use_from_same_package
 // ignore_for_file: implementation_imports
 // ignore_for_file: invalid_use_of_visible_for_testing_member
 // ignore_for_file: prefer_const_constructors
@@ -168,6 +171,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void saveLayer(
     _i2.Rect? bounds,
@@ -183,6 +187,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void restore() => super.noSuchMethod(
         Invocation.method(
@@ -191,6 +196,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void restoreToCount(int? count) => super.noSuchMethod(
         Invocation.method(
@@ -199,6 +205,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   int getSaveCount() => (super.noSuchMethod(
         Invocation.method(
@@ -207,6 +214,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValue: 0,
       ) as int);
+
   @override
   void translate(
     double? dx,
@@ -222,6 +230,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void scale(
     double? sx, [
@@ -237,6 +246,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void rotate(double? radians) => super.noSuchMethod(
         Invocation.method(
@@ -245,6 +255,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void skew(
     double? sx,
@@ -260,6 +271,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void transform(_i5.Float64List? matrix4) => super.noSuchMethod(
         Invocation.method(
@@ -268,6 +280,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i5.Float64List getTransform() => (super.noSuchMethod(
         Invocation.method(
@@ -276,6 +289,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValue: _i5.Float64List(0),
       ) as _i5.Float64List);
+
   @override
   void clipRect(
     _i2.Rect? rect, {
@@ -293,6 +307,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipRRect(
     _i2.RRect? rrect, {
@@ -306,6 +321,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipPath(
     _i2.Path? path, {
@@ -319,6 +335,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i2.Rect getLocalClipBounds() => (super.noSuchMethod(
         Invocation.method(
@@ -333,6 +350,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
           ),
         ),
       ) as _i2.Rect);
+
   @override
   _i2.Rect getDestinationClipBounds() => (super.noSuchMethod(
         Invocation.method(
@@ -347,6 +365,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
           ),
         ),
       ) as _i2.Rect);
+
   @override
   void drawColor(
     _i2.Color? color,
@@ -362,6 +381,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawLine(
     _i2.Offset? p1,
@@ -379,6 +399,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPaint(_i2.Paint? paint) => super.noSuchMethod(
         Invocation.method(
@@ -387,6 +408,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRect(
     _i2.Rect? rect,
@@ -402,6 +424,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRRect(
     _i2.RRect? rrect,
@@ -417,6 +440,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawDRRect(
     _i2.RRect? outer,
@@ -434,6 +458,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawOval(
     _i2.Rect? rect,
@@ -449,6 +474,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawCircle(
     _i2.Offset? c,
@@ -466,6 +492,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawArc(
     _i2.Rect? rect,
@@ -487,6 +514,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPath(
     _i2.Path? path,
@@ -502,6 +530,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawImage(
     _i2.Image? image,
@@ -519,6 +548,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawImageRect(
     _i2.Image? image,
@@ -538,6 +568,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawImageNine(
     _i2.Image? image,
@@ -557,6 +588,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPicture(_i2.Picture? picture) => super.noSuchMethod(
         Invocation.method(
@@ -565,6 +597,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawParagraph(
     _i2.Paragraph? paragraph,
@@ -580,6 +613,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPoints(
     _i2.PointMode? pointMode,
@@ -597,6 +631,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRawPoints(
     _i2.PointMode? pointMode,
@@ -614,6 +649,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawVertices(
     _i2.Vertices? vertices,
@@ -631,6 +667,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawAtlas(
     _i2.Image? atlas,
@@ -656,6 +693,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRawAtlas(
     _i2.Image? atlas,
@@ -681,6 +719,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawShadow(
     _i2.Path? path,
@@ -718,6 +757,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
           Invocation.getter(#canvas),
         ),
       ) as _i2.Canvas);
+
   @override
   _i2.Size get size => (super.noSuchMethod(
         Invocation.getter(#size),
@@ -726,6 +766,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
           Invocation.getter(#size),
         ),
       ) as _i2.Size);
+
   @override
   void drawRRect(
     _i2.RRect? rrect,
@@ -741,6 +782,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void save() => super.noSuchMethod(
         Invocation.method(
@@ -749,6 +791,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void restore() => super.noSuchMethod(
         Invocation.method(
@@ -757,6 +800,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipRect(
     _i2.Rect? rect, {
@@ -774,6 +818,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void translate(
     double? dx,
@@ -789,6 +834,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void rotate(double? radius) => super.noSuchMethod(
         Invocation.method(
@@ -797,6 +843,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPath(
     _i2.Path? path,
@@ -812,6 +859,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void saveLayer(
     _i2.Rect? bounds,
@@ -827,6 +875,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPicture(_i2.Picture? picture) => super.noSuchMethod(
         Invocation.method(
@@ -835,6 +884,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawImage(
     _i2.Image? image,
@@ -852,6 +902,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipPath(
     _i2.Path? path, {
@@ -865,6 +916,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRect(
     _i2.Rect? rect,
@@ -880,6 +932,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawLine(
     _i2.Offset? p1,
@@ -897,6 +950,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawCircle(
     _i2.Offset? center,
@@ -914,6 +968,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawArc(
     _i2.Rect? rect,
@@ -935,6 +990,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawText(
     _i3.TextPainter? tp,
@@ -952,6 +1008,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawDot(
     _i7.FlDotPainter? painter,
@@ -969,6 +1026,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRotated({
     required _i2.Size? size,
@@ -991,6 +1049,7 @@ class MockCanvasWrapper extends _i1.Mock implements _i6.CanvasWrapper {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawDashedLine(
     _i2.Offset? from,
@@ -1028,16 +1087,19 @@ class MockBuildContext extends _i1.Mock implements _i3.BuildContext {
           Invocation.getter(#widget),
         ),
       ) as _i3.Widget);
+
   @override
   bool get mounted => (super.noSuchMethod(
         Invocation.getter(#mounted),
         returnValue: false,
       ) as bool);
+
   @override
   bool get debugDoingBuild => (super.noSuchMethod(
         Invocation.getter(#debugDoingBuild),
         returnValue: false,
       ) as bool);
+
   @override
   _i3.InheritedWidget dependOnInheritedElement(
     _i3.InheritedElement? ancestor, {
@@ -1058,6 +1120,7 @@ class MockBuildContext extends _i1.Mock implements _i3.BuildContext {
           ),
         ),
       ) as _i3.InheritedWidget);
+
   @override
   void visitAncestorElements(_i3.ConditionalElementVisitor? visitor) =>
       super.noSuchMethod(
@@ -1067,6 +1130,7 @@ class MockBuildContext extends _i1.Mock implements _i3.BuildContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void visitChildElements(_i3.ElementVisitor? visitor) => super.noSuchMethod(
         Invocation.method(
@@ -1075,6 +1139,7 @@ class MockBuildContext extends _i1.Mock implements _i3.BuildContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void dispatchNotification(_i3.Notification? notification) =>
       super.noSuchMethod(
@@ -1084,6 +1149,7 @@ class MockBuildContext extends _i1.Mock implements _i3.BuildContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i3.DiagnosticsNode describeElement(
     String? name, {
@@ -1104,6 +1170,7 @@ class MockBuildContext extends _i1.Mock implements _i3.BuildContext {
           ),
         ),
       ) as _i3.DiagnosticsNode);
+
   @override
   _i3.DiagnosticsNode describeWidget(
     String? name, {
@@ -1124,6 +1191,7 @@ class MockBuildContext extends _i1.Mock implements _i3.BuildContext {
           ),
         ),
       ) as _i3.DiagnosticsNode);
+
   @override
   List<_i3.DiagnosticsNode> describeMissingAncestor(
           {required Type? expectedAncestorType}) =>
@@ -1135,6 +1203,7 @@ class MockBuildContext extends _i1.Mock implements _i3.BuildContext {
         ),
         returnValue: <_i3.DiagnosticsNode>[],
       ) as List<_i3.DiagnosticsNode>);
+
   @override
   _i3.DiagnosticsNode describeOwnershipChain(String? name) =>
       (super.noSuchMethod(
@@ -1168,6 +1237,7 @@ class MockUtils extends _i1.Mock implements _i8.Utils {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   double degrees(double? radians) => (super.noSuchMethod(
         Invocation.method(
@@ -1176,6 +1246,7 @@ class MockUtils extends _i1.Mock implements _i8.Utils {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   _i2.Size getDefaultSize(_i2.Size? screenSize) => (super.noSuchMethod(
         Invocation.method(
@@ -1190,6 +1261,7 @@ class MockUtils extends _i1.Mock implements _i8.Utils {
           ),
         ),
       ) as _i2.Size);
+
   @override
   double translateRotatedPosition(
     double? size,
@@ -1205,6 +1277,7 @@ class MockUtils extends _i1.Mock implements _i8.Utils {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   _i2.Offset calculateRotationOffset(
     _i2.Size? size,
@@ -1229,6 +1302,7 @@ class MockUtils extends _i1.Mock implements _i8.Utils {
           ),
         ),
       ) as _i2.Offset);
+
   @override
   _i3.BorderRadius? normalizeBorderRadius(
     _i3.BorderRadius? borderRadius,
@@ -1241,6 +1315,7 @@ class MockUtils extends _i1.Mock implements _i8.Utils {
           width,
         ],
       )) as _i3.BorderRadius?);
+
   @override
   _i3.BorderSide normalizeBorderSide(
     _i3.BorderSide? borderSide,
@@ -1265,6 +1340,7 @@ class MockUtils extends _i1.Mock implements _i8.Utils {
           ),
         ),
       ) as _i3.BorderSide);
+
   @override
   double getEfficientInterval(
     double? axisViewSize,
@@ -1282,6 +1358,7 @@ class MockUtils extends _i1.Mock implements _i8.Utils {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   double roundInterval(double? input) => (super.noSuchMethod(
         Invocation.method(
@@ -1290,14 +1367,44 @@ class MockUtils extends _i1.Mock implements _i8.Utils {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
-  String formatNumber(double? number) => (super.noSuchMethod(
+  int getFractionDigits(double? value) => (super.noSuchMethod(
+        Invocation.method(
+          #getFractionDigits,
+          [value],
+        ),
+        returnValue: 0,
+      ) as int);
+
+  @override
+  String formatNumber(
+    double? axisMin,
+    double? axisMax,
+    double? axisValue,
+  ) =>
+      (super.noSuchMethod(
         Invocation.method(
           #formatNumber,
-          [number],
+          [
+            axisMin,
+            axisMax,
+            axisValue,
+          ],
         ),
-        returnValue: '',
+        returnValue: _i9.dummyValue<String>(
+          this,
+          Invocation.method(
+            #formatNumber,
+            [
+              axisMin,
+              axisMax,
+              axisValue,
+            ],
+          ),
+        ),
       ) as String);
+
   @override
   _i3.TextStyle getThemeAwareTextStyle(
     _i3.BuildContext? context,
@@ -1322,6 +1429,7 @@ class MockUtils extends _i1.Mock implements _i8.Utils {
           ),
         ),
       ) as _i3.TextStyle);
+
   @override
   double getBestInitialIntervalValue(
     double? min,
@@ -1341,6 +1449,7 @@ class MockUtils extends _i1.Mock implements _i8.Utils {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   double convertRadiusToSigma(double? radius) => (super.noSuchMethod(
         Invocation.method(
@@ -1354,7 +1463,7 @@ class MockUtils extends _i1.Mock implements _i8.Utils {
 /// A class which mocks [LineChartPainter].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
+class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
   MockLineChartPainter() {
     _i1.throwOnMissingStub(this);
   }
@@ -1363,7 +1472,7 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
   void paint(
     _i3.BuildContext? context,
     _i6.CanvasWrapper? canvasWrapper,
-    _i10.PaintHolder<_i7.LineChartData>? holder,
+    _i11.PaintHolder<_i7.LineChartData>? holder,
   ) =>
       super.noSuchMethod(
         Invocation.method(
@@ -1376,10 +1485,11 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipToBorder(
     _i6.CanvasWrapper? canvasWrapper,
-    _i10.PaintHolder<_i7.LineChartData>? holder,
+    _i11.PaintHolder<_i7.LineChartData>? holder,
   ) =>
       super.noSuchMethod(
         Invocation.method(
@@ -1391,11 +1501,12 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawBarLine(
     _i6.CanvasWrapper? canvasWrapper,
     _i7.LineChartBarData? barData,
-    _i10.PaintHolder<_i7.LineChartData>? holder,
+    _i11.PaintHolder<_i7.LineChartData>? holder,
   ) =>
       super.noSuchMethod(
         Invocation.method(
@@ -1408,12 +1519,13 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawBetweenBarsArea(
     _i6.CanvasWrapper? canvasWrapper,
     _i7.LineChartData? data,
     _i7.BetweenBarsData? betweenBarsData,
-    _i10.PaintHolder<_i7.LineChartData>? holder,
+    _i11.PaintHolder<_i7.LineChartData>? holder,
   ) =>
       super.noSuchMethod(
         Invocation.method(
@@ -1427,11 +1539,12 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawDots(
     _i6.CanvasWrapper? canvasWrapper,
     _i7.LineChartBarData? barData,
-    _i10.PaintHolder<_i7.LineChartData>? holder,
+    _i11.PaintHolder<_i7.LineChartData>? holder,
   ) =>
       super.noSuchMethod(
         Invocation.method(
@@ -1444,11 +1557,12 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawTouchedSpotsIndicator(
     _i6.CanvasWrapper? canvasWrapper,
-    List<_i9.LineIndexDrawingInfo>? lineIndexDrawingInfo,
-    _i10.PaintHolder<_i7.LineChartData>? holder,
+    List<_i10.LineIndexDrawingInfo>? lineIndexDrawingInfo,
+    _i11.PaintHolder<_i7.LineChartData>? holder,
   ) =>
       super.noSuchMethod(
         Invocation.method(
@@ -1461,12 +1575,13 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i2.Path generateBarPath(
     _i2.Size? viewSize,
     _i7.LineChartBarData? barData,
     List<_i7.FlSpot>? barSpots,
-    _i10.PaintHolder<_i7.LineChartData>? holder, {
+    _i11.PaintHolder<_i7.LineChartData>? holder, {
     _i2.Path? appendToPath,
   }) =>
       (super.noSuchMethod(
@@ -1494,12 +1609,13 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
           ),
         ),
       ) as _i2.Path);
+
   @override
   _i2.Path generateNormalBarPath(
     _i2.Size? viewSize,
     _i7.LineChartBarData? barData,
     List<_i7.FlSpot>? barSpots,
-    _i10.PaintHolder<_i7.LineChartData>? holder, {
+    _i11.PaintHolder<_i7.LineChartData>? holder, {
     _i2.Path? appendToPath,
   }) =>
       (super.noSuchMethod(
@@ -1527,12 +1643,13 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
           ),
         ),
       ) as _i2.Path);
+
   @override
   _i2.Path generateStepBarPath(
     _i2.Size? viewSize,
     _i7.LineChartBarData? barData,
     List<_i7.FlSpot>? barSpots,
-    _i10.PaintHolder<_i7.LineChartData>? holder, {
+    _i11.PaintHolder<_i7.LineChartData>? holder, {
     _i2.Path? appendToPath,
   }) =>
       (super.noSuchMethod(
@@ -1560,13 +1677,14 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
           ),
         ),
       ) as _i2.Path);
+
   @override
   _i2.Path generateBelowBarPath(
     _i2.Size? viewSize,
     _i7.LineChartBarData? barData,
     _i2.Path? barPath,
     List<_i7.FlSpot>? barSpots,
-    _i10.PaintHolder<_i7.LineChartData>? holder, {
+    _i11.PaintHolder<_i7.LineChartData>? holder, {
     bool? fillCompletely = false,
   }) =>
       (super.noSuchMethod(
@@ -1596,13 +1714,14 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
           ),
         ),
       ) as _i2.Path);
+
   @override
   _i2.Path generateAboveBarPath(
     _i2.Size? viewSize,
     _i7.LineChartBarData? barData,
     _i2.Path? barPath,
     List<_i7.FlSpot>? barSpots,
-    _i10.PaintHolder<_i7.LineChartData>? holder, {
+    _i11.PaintHolder<_i7.LineChartData>? holder, {
     bool? fillCompletely = false,
   }) =>
       (super.noSuchMethod(
@@ -1632,12 +1751,13 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
           ),
         ),
       ) as _i2.Path);
+
   @override
   void drawBelowBar(
     _i6.CanvasWrapper? canvasWrapper,
     _i2.Path? belowBarPath,
     _i2.Path? filledAboveBarPath,
-    _i10.PaintHolder<_i7.LineChartData>? holder,
+    _i11.PaintHolder<_i7.LineChartData>? holder,
     _i7.LineChartBarData? barData,
   ) =>
       super.noSuchMethod(
@@ -1653,12 +1773,13 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawAboveBar(
     _i6.CanvasWrapper? canvasWrapper,
     _i2.Path? aboveBarPath,
     _i2.Path? filledBelowBarPath,
-    _i10.PaintHolder<_i7.LineChartData>? holder,
+    _i11.PaintHolder<_i7.LineChartData>? holder,
     _i7.LineChartBarData? barData,
   ) =>
       super.noSuchMethod(
@@ -1674,13 +1795,14 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawBetweenBar(
     _i6.CanvasWrapper? canvasWrapper,
     _i2.Path? barPath,
     _i7.BetweenBarsData? betweenBarsData,
     _i2.Rect? aroundRect,
-    _i10.PaintHolder<_i7.LineChartData>? holder,
+    _i11.PaintHolder<_i7.LineChartData>? holder,
   ) =>
       super.noSuchMethod(
         Invocation.method(
@@ -1695,6 +1817,7 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawBarShadow(
     _i6.CanvasWrapper? canvasWrapper,
@@ -1712,12 +1835,13 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawBar(
     _i6.CanvasWrapper? canvasWrapper,
     _i2.Path? barPath,
     _i7.LineChartBarData? barData,
-    _i10.PaintHolder<_i7.LineChartData>? holder,
+    _i11.PaintHolder<_i7.LineChartData>? holder,
   ) =>
       super.noSuchMethod(
         Invocation.method(
@@ -1731,6 +1855,7 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawTouchTooltip(
     _i3.BuildContext? context,
@@ -1738,7 +1863,7 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
     _i7.LineTouchTooltipData? tooltipData,
     _i7.FlSpot? showOnSpot,
     _i7.ShowingTooltipIndicators? showingTooltipSpots,
-    _i10.PaintHolder<_i7.LineChartData>? holder,
+    _i11.PaintHolder<_i7.LineChartData>? holder,
   ) =>
       super.noSuchMethod(
         Invocation.method(
@@ -1754,11 +1879,12 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   double getBarLineXLength(
     _i7.LineChartBarData? barData,
     _i2.Size? chartUsableSize,
-    _i10.PaintHolder<_i7.LineChartData>? holder,
+    _i11.PaintHolder<_i7.LineChartData>? holder,
   ) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -1771,11 +1897,12 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   List<_i7.TouchLineBarSpot>? handleTouch(
     _i2.Offset? localPosition,
     _i2.Size? size,
-    _i10.PaintHolder<_i7.LineChartData>? holder,
+    _i11.PaintHolder<_i7.LineChartData>? holder,
   ) =>
       (super.noSuchMethod(Invocation.method(
         #handleTouch,
@@ -1785,13 +1912,14 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
           holder,
         ],
       )) as List<_i7.TouchLineBarSpot>?);
+
   @override
   _i7.TouchLineBarSpot? getNearestTouchedSpot(
     _i2.Size? viewSize,
     _i2.Offset? touchedPoint,
     _i7.LineChartBarData? barData,
     int? barDataPosition,
-    _i10.PaintHolder<_i7.LineChartData>? holder,
+    _i11.PaintHolder<_i7.LineChartData>? holder,
   ) =>
       (super.noSuchMethod(Invocation.method(
         #getNearestTouchedSpot,
@@ -1803,10 +1931,11 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
           holder,
         ],
       )) as _i7.TouchLineBarSpot?);
+
   @override
   void drawGrid(
     _i6.CanvasWrapper? canvasWrapper,
-    _i10.PaintHolder<_i7.LineChartData>? holder,
+    _i11.PaintHolder<_i7.LineChartData>? holder,
   ) =>
       super.noSuchMethod(
         Invocation.method(
@@ -1818,10 +1947,11 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawBackground(
     _i6.CanvasWrapper? canvasWrapper,
-    _i10.PaintHolder<_i7.LineChartData>? holder,
+    _i11.PaintHolder<_i7.LineChartData>? holder,
   ) =>
       super.noSuchMethod(
         Invocation.method(
@@ -1833,10 +1963,11 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRangeAnnotation(
     _i6.CanvasWrapper? canvasWrapper,
-    _i10.PaintHolder<_i7.LineChartData>? holder,
+    _i11.PaintHolder<_i7.LineChartData>? holder,
   ) =>
       super.noSuchMethod(
         Invocation.method(
@@ -1848,11 +1979,12 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawExtraLines(
     _i3.BuildContext? context,
     _i6.CanvasWrapper? canvasWrapper,
-    _i10.PaintHolder<_i7.LineChartData>? holder,
+    _i11.PaintHolder<_i7.LineChartData>? holder,
   ) =>
       super.noSuchMethod(
         Invocation.method(
@@ -1865,11 +1997,12 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawHorizontalLines(
     _i3.BuildContext? context,
     _i6.CanvasWrapper? canvasWrapper,
-    _i10.PaintHolder<_i7.LineChartData>? holder,
+    _i11.PaintHolder<_i7.LineChartData>? holder,
     _i2.Size? viewSize,
   ) =>
       super.noSuchMethod(
@@ -1884,11 +2017,12 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawVerticalLines(
     _i3.BuildContext? context,
     _i6.CanvasWrapper? canvasWrapper,
-    _i10.PaintHolder<_i7.LineChartData>? holder,
+    _i11.PaintHolder<_i7.LineChartData>? holder,
     _i2.Size? viewSize,
   ) =>
       super.noSuchMethod(
@@ -1903,11 +2037,12 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   double getPixelX(
     double? spotX,
     _i2.Size? viewSize,
-    _i10.PaintHolder<_i7.LineChartData>? holder,
+    _i11.PaintHolder<_i7.LineChartData>? holder,
   ) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -1920,11 +2055,12 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   double getPixelY(
     double? spotY,
     _i2.Size? viewSize,
-    _i10.PaintHolder<_i7.LineChartData>? holder,
+    _i11.PaintHolder<_i7.LineChartData>? holder,
   ) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -1937,6 +2073,7 @@ class MockLineChartPainter extends _i1.Mock implements _i9.LineChartPainter {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   double getTooltipLeft(
     double? dx,
