@@ -1,4 +1,4 @@
-// Mocks generated by Mockito 5.4.0 from annotations
+// Mocks generated by Mockito 5.4.3 from annotations
 // in fl_chart/test/chart/line_chart/line_chart_renderer_test.dart.
 // Do not manually edit this file.
 
@@ -23,6 +23,8 @@ import 'package:mockito/mockito.dart' as _i1;
 // ignore_for_file: avoid_redundant_argument_values
 // ignore_for_file: avoid_setters_without_getters
 // ignore_for_file: comment_references
+// ignore_for_file: deprecated_member_use
+// ignore_for_file: deprecated_member_use_from_same_package
 // ignore_for_file: implementation_imports
 // ignore_for_file: invalid_use_of_visible_for_testing_member
 // ignore_for_file: prefer_const_constructors
@@ -163,6 +165,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void saveLayer(
     _i2.Rect? bounds,
@@ -178,6 +181,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void restore() => super.noSuchMethod(
         Invocation.method(
@@ -186,6 +190,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void restoreToCount(int? count) => super.noSuchMethod(
         Invocation.method(
@@ -194,6 +199,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   int getSaveCount() => (super.noSuchMethod(
         Invocation.method(
@@ -202,6 +208,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValue: 0,
       ) as int);
+
   @override
   void translate(
     double? dx,
@@ -217,6 +224,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void scale(
     double? sx, [
@@ -232,6 +240,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void rotate(double? radians) => super.noSuchMethod(
         Invocation.method(
@@ -240,6 +249,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void skew(
     double? sx,
@@ -255,6 +265,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void transform(_i7.Float64List? matrix4) => super.noSuchMethod(
         Invocation.method(
@@ -263,6 +274,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i7.Float64List getTransform() => (super.noSuchMethod(
         Invocation.method(
@@ -271,6 +283,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValue: _i7.Float64List(0),
       ) as _i7.Float64List);
+
   @override
   void clipRect(
     _i2.Rect? rect, {
@@ -288,6 +301,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipRRect(
     _i2.RRect? rrect, {
@@ -301,6 +315,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipPath(
     _i2.Path? path, {
@@ -314,6 +329,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i2.Rect getLocalClipBounds() => (super.noSuchMethod(
         Invocation.method(
@@ -328,6 +344,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
           ),
         ),
       ) as _i2.Rect);
+
   @override
   _i2.Rect getDestinationClipBounds() => (super.noSuchMethod(
         Invocation.method(
@@ -342,6 +359,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
           ),
         ),
       ) as _i2.Rect);
+
   @override
   void drawColor(
     _i2.Color? color,
@@ -357,6 +375,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawLine(
     _i2.Offset? p1,
@@ -374,6 +393,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPaint(_i2.Paint? paint) => super.noSuchMethod(
         Invocation.method(
@@ -382,6 +402,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRect(
     _i2.Rect? rect,
@@ -397,6 +418,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRRect(
     _i2.RRect? rrect,
@@ -412,6 +434,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawDRRect(
     _i2.RRect? outer,
@@ -429,6 +452,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawOval(
     _i2.Rect? rect,
@@ -444,6 +468,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawCircle(
     _i2.Offset? c,
@@ -461,6 +486,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawArc(
     _i2.Rect? rect,
@@ -482,6 +508,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPath(
     _i2.Path? path,
@@ -497,6 +524,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawImage(
     _i2.Image? image,
@@ -514,6 +542,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawImageRect(
     _i2.Image? image,
@@ -533,6 +562,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawImageNine(
     _i2.Image? image,
@@ -552,6 +582,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPicture(_i2.Picture? picture) => super.noSuchMethod(
         Invocation.method(
@@ -560,6 +591,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawParagraph(
     _i2.Paragraph? paragraph,
@@ -575,6 +607,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPoints(
     _i2.PointMode? pointMode,
@@ -592,6 +625,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRawPoints(
     _i2.PointMode? pointMode,
@@ -609,6 +643,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawVertices(
     _i2.Vertices? vertices,
@@ -626,6 +661,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawAtlas(
     _i2.Image? atlas,
@@ -651,6 +687,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRawAtlas(
     _i2.Image? atlas,
@@ -676,6 +713,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawShadow(
     _i2.Path? path,
@@ -713,6 +751,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           Invocation.getter(#estimatedBounds),
         ),
       ) as _i2.Rect);
+
   @override
   _i2.Canvas get canvas => (super.noSuchMethod(
         Invocation.getter(#canvas),
@@ -721,6 +760,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           Invocation.getter(#canvas),
         ),
       ) as _i2.Canvas);
+
   @override
   void paintChild(
     _i3.RenderObject? child,
@@ -736,6 +776,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void appendLayer(_i4.Layer? layer) => super.noSuchMethod(
         Invocation.method(
@@ -744,6 +785,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i2.VoidCallback addCompositionCallback(_i4.CompositionCallback? callback) =>
       (super.noSuchMethod(
@@ -753,6 +795,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValue: () {},
       ) as _i2.VoidCallback);
+
   @override
   void stopRecordingIfNeeded() => super.noSuchMethod(
         Invocation.method(
@@ -761,6 +804,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void setIsComplexHint() => super.noSuchMethod(
         Invocation.method(
@@ -769,6 +813,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void setWillChangeHint() => super.noSuchMethod(
         Invocation.method(
@@ -777,6 +822,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void addLayer(_i4.Layer? layer) => super.noSuchMethod(
         Invocation.method(
@@ -785,6 +831,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void pushLayer(
     _i4.ContainerLayer? childLayer,
@@ -804,6 +851,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i3.PaintingContext createChildContext(
     _i4.ContainerLayer? childLayer,
@@ -828,6 +876,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           ),
         ),
       ) as _i3.PaintingContext);
+
   @override
   _i4.ClipRectLayer? pushClipRect(
     bool? needsCompositing,
@@ -850,6 +899,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           #oldLayer: oldLayer,
         },
       )) as _i4.ClipRectLayer?);
+
   @override
   _i4.ClipRRectLayer? pushClipRRect(
     bool? needsCompositing,
@@ -874,6 +924,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           #oldLayer: oldLayer,
         },
       )) as _i4.ClipRRectLayer?);
+
   @override
   _i4.ClipPathLayer? pushClipPath(
     bool? needsCompositing,
@@ -898,6 +949,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           #oldLayer: oldLayer,
         },
       )) as _i4.ClipPathLayer?);
+
   @override
   _i4.ColorFilterLayer pushColorFilter(
     _i2.Offset? offset,
@@ -928,6 +980,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           ),
         ),
       ) as _i4.ColorFilterLayer);
+
   @override
   _i4.TransformLayer? pushTransform(
     bool? needsCompositing,
@@ -946,6 +999,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ],
         {#oldLayer: oldLayer},
       )) as _i4.TransformLayer?);
+
   @override
   _i4.OpacityLayer pushOpacity(
     _i2.Offset? offset,
@@ -976,6 +1030,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           ),
         ),
       ) as _i4.OpacityLayer);
+
   @override
   void clipPathAndPaint(
     _i2.Path? path,
@@ -995,6 +1050,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipRRectAndPaint(
     _i2.RRect? rrect,
@@ -1014,6 +1070,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipRectAndPaint(
     _i2.Rect? rect,
@@ -1051,16 +1108,19 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
           Invocation.getter(#widget),
         ),
       ) as _i6.Widget);
+
   @override
   bool get mounted => (super.noSuchMethod(
         Invocation.getter(#mounted),
         returnValue: false,
       ) as bool);
+
   @override
   bool get debugDoingBuild => (super.noSuchMethod(
         Invocation.getter(#debugDoingBuild),
         returnValue: false,
       ) as bool);
+
   @override
   _i6.InheritedWidget dependOnInheritedElement(
     _i6.InheritedElement? ancestor, {
@@ -1081,6 +1141,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
           ),
         ),
       ) as _i6.InheritedWidget);
+
   @override
   void visitAncestorElements(_i6.ConditionalElementVisitor? visitor) =>
       super.noSuchMethod(
@@ -1090,6 +1151,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void visitChildElements(_i6.ElementVisitor? visitor) => super.noSuchMethod(
         Invocation.method(
@@ -1098,6 +1160,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void dispatchNotification(_i9.Notification? notification) =>
       super.noSuchMethod(
@@ -1107,6 +1170,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i5.DiagnosticsNode describeElement(
     String? name, {
@@ -1127,6 +1191,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
           ),
         ),
       ) as _i5.DiagnosticsNode);
+
   @override
   _i5.DiagnosticsNode describeWidget(
     String? name, {
@@ -1147,6 +1212,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
           ),
         ),
       ) as _i5.DiagnosticsNode);
+
   @override
   List<_i5.DiagnosticsNode> describeMissingAncestor(
           {required Type? expectedAncestorType}) =>
@@ -1158,6 +1224,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
         ),
         returnValue: <_i5.DiagnosticsNode>[],
       ) as List<_i5.DiagnosticsNode>);
+
   @override
   _i5.DiagnosticsNode describeOwnershipChain(String? name) =>
       (super.noSuchMethod(
@@ -1200,6 +1267,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipToBorder(
     _i11.CanvasWrapper? canvasWrapper,
@@ -1215,6 +1283,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawBarLine(
     _i11.CanvasWrapper? canvasWrapper,
@@ -1232,6 +1301,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawBetweenBarsArea(
     _i11.CanvasWrapper? canvasWrapper,
@@ -1251,6 +1321,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawDots(
     _i11.CanvasWrapper? canvasWrapper,
@@ -1268,6 +1339,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawTouchedSpotsIndicator(
     _i11.CanvasWrapper? canvasWrapper,
@@ -1285,6 +1357,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i2.Path generateBarPath(
     _i2.Size? viewSize,
@@ -1318,6 +1391,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
           ),
         ),
       ) as _i2.Path);
+
   @override
   _i2.Path generateNormalBarPath(
     _i2.Size? viewSize,
@@ -1351,6 +1425,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
           ),
         ),
       ) as _i2.Path);
+
   @override
   _i2.Path generateStepBarPath(
     _i2.Size? viewSize,
@@ -1384,6 +1459,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
           ),
         ),
       ) as _i2.Path);
+
   @override
   _i2.Path generateBelowBarPath(
     _i2.Size? viewSize,
@@ -1420,6 +1496,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
           ),
         ),
       ) as _i2.Path);
+
   @override
   _i2.Path generateAboveBarPath(
     _i2.Size? viewSize,
@@ -1456,6 +1533,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
           ),
         ),
       ) as _i2.Path);
+
   @override
   void drawBelowBar(
     _i11.CanvasWrapper? canvasWrapper,
@@ -1477,6 +1555,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawAboveBar(
     _i11.CanvasWrapper? canvasWrapper,
@@ -1498,6 +1577,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawBetweenBar(
     _i11.CanvasWrapper? canvasWrapper,
@@ -1519,6 +1599,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawBarShadow(
     _i11.CanvasWrapper? canvasWrapper,
@@ -1536,6 +1617,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawBar(
     _i11.CanvasWrapper? canvasWrapper,
@@ -1555,6 +1637,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawTouchTooltip(
     _i6.BuildContext? context,
@@ -1578,6 +1661,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   double getBarLineXLength(
     _i13.LineChartBarData? barData,
@@ -1595,6 +1679,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   List<_i13.TouchLineBarSpot>? handleTouch(
     _i2.Offset? localPosition,
@@ -1609,6 +1694,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
           holder,
         ],
       )) as List<_i13.TouchLineBarSpot>?);
+
   @override
   _i13.TouchLineBarSpot? getNearestTouchedSpot(
     _i2.Size? viewSize,
@@ -1627,6 +1713,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
           holder,
         ],
       )) as _i13.TouchLineBarSpot?);
+
   @override
   void drawGrid(
     _i11.CanvasWrapper? canvasWrapper,
@@ -1642,6 +1729,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawBackground(
     _i11.CanvasWrapper? canvasWrapper,
@@ -1657,6 +1745,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRangeAnnotation(
     _i11.CanvasWrapper? canvasWrapper,
@@ -1672,6 +1761,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawExtraLines(
     _i6.BuildContext? context,
@@ -1689,6 +1779,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawHorizontalLines(
     _i6.BuildContext? context,
@@ -1708,6 +1799,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawVerticalLines(
     _i6.BuildContext? context,
@@ -1727,6 +1819,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   double getPixelX(
     double? spotX,
@@ -1744,6 +1837,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   double getPixelY(
     double? spotY,
@@ -1761,6 +1855,7 @@ class MockLineChartPainter extends _i1.Mock implements _i10.LineChartPainter {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   double getTooltipLeft(
     double? dx,
