@@ -1,4 +1,4 @@
-// Mocks generated by Mockito 5.4.0 from annotations
+// Mocks generated by Mockito 5.4.3 from annotations
 // in fl_chart/test/chart/pie_chart/pie_chart_renderer_test.dart.
 // Do not manually edit this file.
 
@@ -24,6 +24,8 @@ import 'package:mockito/mockito.dart' as _i1;
 // ignore_for_file: avoid_redundant_argument_values
 // ignore_for_file: avoid_setters_without_getters
 // ignore_for_file: comment_references
+// ignore_for_file: deprecated_member_use
+// ignore_for_file: deprecated_member_use_from_same_package
 // ignore_for_file: implementation_imports
 // ignore_for_file: invalid_use_of_visible_for_testing_member
 // ignore_for_file: prefer_const_constructors
@@ -175,6 +177,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void saveLayer(
     _i2.Rect? bounds,
@@ -190,6 +193,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void restore() => super.noSuchMethod(
         Invocation.method(
@@ -198,6 +202,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void restoreToCount(int? count) => super.noSuchMethod(
         Invocation.method(
@@ -206,6 +211,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   int getSaveCount() => (super.noSuchMethod(
         Invocation.method(
@@ -214,6 +220,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValue: 0,
       ) as int);
+
   @override
   void translate(
     double? dx,
@@ -229,6 +236,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void scale(
     double? sx, [
@@ -244,6 +252,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void rotate(double? radians) => super.noSuchMethod(
         Invocation.method(
@@ -252,6 +261,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void skew(
     double? sx,
@@ -267,6 +277,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void transform(_i8.Float64List? matrix4) => super.noSuchMethod(
         Invocation.method(
@@ -275,6 +286,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i8.Float64List getTransform() => (super.noSuchMethod(
         Invocation.method(
@@ -283,6 +295,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValue: _i8.Float64List(0),
       ) as _i8.Float64List);
+
   @override
   void clipRect(
     _i2.Rect? rect, {
@@ -300,6 +313,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipRRect(
     _i2.RRect? rrect, {
@@ -313,6 +327,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipPath(
     _i2.Path? path, {
@@ -326,6 +341,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i2.Rect getLocalClipBounds() => (super.noSuchMethod(
         Invocation.method(
@@ -340,6 +356,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
           ),
         ),
       ) as _i2.Rect);
+
   @override
   _i2.Rect getDestinationClipBounds() => (super.noSuchMethod(
         Invocation.method(
@@ -354,6 +371,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
           ),
         ),
       ) as _i2.Rect);
+
   @override
   void drawColor(
     _i2.Color? color,
@@ -369,6 +387,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawLine(
     _i2.Offset? p1,
@@ -386,6 +405,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPaint(_i2.Paint? paint) => super.noSuchMethod(
         Invocation.method(
@@ -394,6 +414,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRect(
     _i2.Rect? rect,
@@ -409,6 +430,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRRect(
     _i2.RRect? rrect,
@@ -424,6 +446,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawDRRect(
     _i2.RRect? outer,
@@ -441,6 +464,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawOval(
     _i2.Rect? rect,
@@ -456,6 +480,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawCircle(
     _i2.Offset? c,
@@ -473,6 +498,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawArc(
     _i2.Rect? rect,
@@ -494,6 +520,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPath(
     _i2.Path? path,
@@ -509,6 +536,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawImage(
     _i2.Image? image,
@@ -526,6 +554,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawImageRect(
     _i2.Image? image,
@@ -545,6 +574,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawImageNine(
     _i2.Image? image,
@@ -564,6 +594,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPicture(_i2.Picture? picture) => super.noSuchMethod(
         Invocation.method(
@@ -572,6 +603,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawParagraph(
     _i2.Paragraph? paragraph,
@@ -587,6 +619,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPoints(
     _i2.PointMode? pointMode,
@@ -604,6 +637,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRawPoints(
     _i2.PointMode? pointMode,
@@ -621,6 +655,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawVertices(
     _i2.Vertices? vertices,
@@ -638,6 +673,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawAtlas(
     _i2.Image? atlas,
@@ -663,6 +699,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRawAtlas(
     _i2.Image? atlas,
@@ -688,6 +725,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawShadow(
     _i2.Path? path,
@@ -725,6 +763,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           Invocation.getter(#estimatedBounds),
         ),
       ) as _i2.Rect);
+
   @override
   _i2.Canvas get canvas => (super.noSuchMethod(
         Invocation.getter(#canvas),
@@ -733,6 +772,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           Invocation.getter(#canvas),
         ),
       ) as _i2.Canvas);
+
   @override
   void paintChild(
     _i3.RenderObject? child,
@@ -748,6 +788,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void appendLayer(_i4.Layer? layer) => super.noSuchMethod(
         Invocation.method(
@@ -756,6 +797,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i2.VoidCallback addCompositionCallback(_i4.CompositionCallback? callback) =>
       (super.noSuchMethod(
@@ -765,6 +807,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValue: () {},
       ) as _i2.VoidCallback);
+
   @override
   void stopRecordingIfNeeded() => super.noSuchMethod(
         Invocation.method(
@@ -773,6 +816,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void setIsComplexHint() => super.noSuchMethod(
         Invocation.method(
@@ -781,6 +825,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void setWillChangeHint() => super.noSuchMethod(
         Invocation.method(
@@ -789,6 +834,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void addLayer(_i4.Layer? layer) => super.noSuchMethod(
         Invocation.method(
@@ -797,6 +843,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void pushLayer(
     _i4.ContainerLayer? childLayer,
@@ -816,6 +863,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i3.PaintingContext createChildContext(
     _i4.ContainerLayer? childLayer,
@@ -840,6 +888,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           ),
         ),
       ) as _i3.PaintingContext);
+
   @override
   _i4.ClipRectLayer? pushClipRect(
     bool? needsCompositing,
@@ -862,6 +911,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           #oldLayer: oldLayer,
         },
       )) as _i4.ClipRectLayer?);
+
   @override
   _i4.ClipRRectLayer? pushClipRRect(
     bool? needsCompositing,
@@ -886,6 +936,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           #oldLayer: oldLayer,
         },
       )) as _i4.ClipRRectLayer?);
+
   @override
   _i4.ClipPathLayer? pushClipPath(
     bool? needsCompositing,
@@ -910,6 +961,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           #oldLayer: oldLayer,
         },
       )) as _i4.ClipPathLayer?);
+
   @override
   _i4.ColorFilterLayer pushColorFilter(
     _i2.Offset? offset,
@@ -940,6 +992,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           ),
         ),
       ) as _i4.ColorFilterLayer);
+
   @override
   _i4.TransformLayer? pushTransform(
     bool? needsCompositing,
@@ -958,6 +1011,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ],
         {#oldLayer: oldLayer},
       )) as _i4.TransformLayer?);
+
   @override
   _i4.OpacityLayer pushOpacity(
     _i2.Offset? offset,
@@ -988,6 +1042,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           ),
         ),
       ) as _i4.OpacityLayer);
+
   @override
   void clipPathAndPaint(
     _i2.Path? path,
@@ -1007,6 +1062,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipRRectAndPaint(
     _i2.RRect? rrect,
@@ -1026,6 +1082,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipRectAndPaint(
     _i2.Rect? rect,
@@ -1063,16 +1120,19 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
           Invocation.getter(#widget),
         ),
       ) as _i6.Widget);
+
   @override
   bool get mounted => (super.noSuchMethod(
         Invocation.getter(#mounted),
         returnValue: false,
       ) as bool);
+
   @override
   bool get debugDoingBuild => (super.noSuchMethod(
         Invocation.getter(#debugDoingBuild),
         returnValue: false,
       ) as bool);
+
   @override
   _i6.InheritedWidget dependOnInheritedElement(
     _i6.InheritedElement? ancestor, {
@@ -1093,6 +1153,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
           ),
         ),
       ) as _i6.InheritedWidget);
+
   @override
   void visitAncestorElements(_i6.ConditionalElementVisitor? visitor) =>
       super.noSuchMethod(
@@ -1102,6 +1163,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void visitChildElements(_i6.ElementVisitor? visitor) => super.noSuchMethod(
         Invocation.method(
@@ -1110,6 +1172,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void dispatchNotification(_i10.Notification? notification) =>
       super.noSuchMethod(
@@ -1119,6 +1182,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i5.DiagnosticsNode describeElement(
     String? name, {
@@ -1139,6 +1203,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
           ),
         ),
       ) as _i5.DiagnosticsNode);
+
   @override
   _i5.DiagnosticsNode describeWidget(
     String? name, {
@@ -1159,6 +1224,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
           ),
         ),
       ) as _i5.DiagnosticsNode);
+
   @override
   List<_i5.DiagnosticsNode> describeMissingAncestor(
           {required Type? expectedAncestorType}) =>
@@ -1170,6 +1236,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
         ),
         returnValue: <_i5.DiagnosticsNode>[],
       ) as List<_i5.DiagnosticsNode>);
+
   @override
   _i5.DiagnosticsNode describeOwnershipChain(String? name) =>
       (super.noSuchMethod(
@@ -1212,6 +1279,7 @@ class MockPieChartPainter extends _i1.Mock implements _i11.PieChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   List<double> calculateSectionsAngle(
     List<_i7.PieChartSectionData>? sections,
@@ -1227,6 +1295,7 @@ class MockPieChartPainter extends _i1.Mock implements _i11.PieChartPainter {
         ),
         returnValue: <double>[],
       ) as List<double>);
+
   @override
   void drawCenterSpace(
     _i12.CanvasWrapper? canvasWrapper,
@@ -1244,6 +1313,7 @@ class MockPieChartPainter extends _i1.Mock implements _i11.PieChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawSections(
     _i12.CanvasWrapper? canvasWrapper,
@@ -1263,6 +1333,7 @@ class MockPieChartPainter extends _i1.Mock implements _i11.PieChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i2.Path generateSectionPath(
     _i7.PieChartSectionData? section,
@@ -1299,6 +1370,7 @@ class MockPieChartPainter extends _i1.Mock implements _i11.PieChartPainter {
           ),
         ),
       ) as _i2.Path);
+
   @override
   _i2.Path createRectPathAroundLine(
     _i14.Line? line,
@@ -1323,6 +1395,7 @@ class MockPieChartPainter extends _i1.Mock implements _i11.PieChartPainter {
           ),
         ),
       ) as _i2.Path);
+
   @override
   void drawSection(
     _i7.PieChartSectionData? section,
@@ -1340,6 +1413,7 @@ class MockPieChartPainter extends _i1.Mock implements _i11.PieChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawSectionStroke(
     _i7.PieChartSectionData? section,
@@ -1359,6 +1433,7 @@ class MockPieChartPainter extends _i1.Mock implements _i11.PieChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawTexts(
     _i6.BuildContext? context,
@@ -1378,6 +1453,7 @@ class MockPieChartPainter extends _i1.Mock implements _i11.PieChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   double calculateCenterRadius(
     _i2.Size? viewSize,
@@ -1393,6 +1469,7 @@ class MockPieChartPainter extends _i1.Mock implements _i11.PieChartPainter {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   _i7.PieTouchedSection handleTouch(
     _i2.Offset? localPosition,
@@ -1420,6 +1497,7 @@ class MockPieChartPainter extends _i1.Mock implements _i11.PieChartPainter {
           ),
         ),
       ) as _i7.PieTouchedSection);
+
   @override
   Map<int, _i2.Offset> getBadgeOffsets(
     _i2.Size? viewSize,
