@@ -24,14 +24,14 @@ void main() {
       tickCount: 1,
     );
 
-    const textScale = 4.0;
+    const textScaler = TextScaler.linear(4);
 
     final mockBuildContext = MockBuildContext();
     final renderRadarChart = RenderRadarChart(
       mockBuildContext,
       data,
       targetData,
-      textScale,
+      textScaler,
     );
 
     final mockPainter = MockRadarChartPainter();
@@ -47,10 +47,10 @@ void main() {
       expect(renderRadarChart.data == data, true);
       expect(renderRadarChart.data == targetData, false);
       expect(renderRadarChart.targetData == targetData, true);
-      expect(renderRadarChart.textScale == textScale, true);
+      expect(renderRadarChart.textScaler == textScaler, true);
       expect(renderRadarChart.paintHolder.data == data, true);
       expect(renderRadarChart.paintHolder.targetData == targetData, true);
-      expect(renderRadarChart.paintHolder.textScale == textScale, true);
+      expect(renderRadarChart.paintHolder.textScaler == textScaler, true);
     });
 
     test('test 2 check paint function', () {
@@ -67,7 +67,7 @@ void main() {
       final paintHolder = result.captured[1] as PaintHolder;
       expect(paintHolder.data, data);
       expect(paintHolder.targetData, targetData);
-      expect(paintHolder.textScale, textScale);
+      expect(paintHolder.textScaler, textScaler);
 
       verify(mockCanvas.restore()).called(1);
     });
@@ -91,18 +91,18 @@ void main() {
       final paintHolder = results[0]['paint_holder'] as PaintHolder;
       expect(paintHolder.data, data);
       expect(paintHolder.targetData, targetData);
-      expect(paintHolder.textScale, textScale);
+      expect(paintHolder.textScaler, textScaler);
     });
 
     test('test 4 check setters', () {
       renderRadarChart
         ..data = targetData
         ..targetData = data
-        ..textScale = 22;
+        ..textScaler = const TextScaler.linear(22);
 
       expect(renderRadarChart.data, targetData);
       expect(renderRadarChart.targetData, data);
-      expect(renderRadarChart.textScale, 22);
+      expect(renderRadarChart.textScaler, const TextScaler.linear(22));
     });
   });
 }
