@@ -1,4 +1,4 @@
-// Mocks generated by Mockito 5.4.0 from annotations
+// Mocks generated by Mockito 5.4.3 from annotations
 // in fl_chart/test/chart/radar_chart/radar_chart_renderer_test.dart.
 // Do not manually edit this file.
 
@@ -24,6 +24,8 @@ import 'package:mockito/mockito.dart' as _i1;
 // ignore_for_file: avoid_redundant_argument_values
 // ignore_for_file: avoid_setters_without_getters
 // ignore_for_file: comment_references
+// ignore_for_file: deprecated_member_use
+// ignore_for_file: deprecated_member_use_from_same_package
 // ignore_for_file: implementation_imports
 // ignore_for_file: invalid_use_of_visible_for_testing_member
 // ignore_for_file: prefer_const_constructors
@@ -154,6 +156,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void saveLayer(
     _i2.Rect? bounds,
@@ -169,6 +172,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void restore() => super.noSuchMethod(
         Invocation.method(
@@ -177,6 +181,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void restoreToCount(int? count) => super.noSuchMethod(
         Invocation.method(
@@ -185,6 +190,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   int getSaveCount() => (super.noSuchMethod(
         Invocation.method(
@@ -193,6 +199,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValue: 0,
       ) as int);
+
   @override
   void translate(
     double? dx,
@@ -208,6 +215,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void scale(
     double? sx, [
@@ -223,6 +231,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void rotate(double? radians) => super.noSuchMethod(
         Invocation.method(
@@ -231,6 +240,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void skew(
     double? sx,
@@ -246,6 +256,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void transform(_i7.Float64List? matrix4) => super.noSuchMethod(
         Invocation.method(
@@ -254,6 +265,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i7.Float64List getTransform() => (super.noSuchMethod(
         Invocation.method(
@@ -262,6 +274,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValue: _i7.Float64List(0),
       ) as _i7.Float64List);
+
   @override
   void clipRect(
     _i2.Rect? rect, {
@@ -279,6 +292,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipRRect(
     _i2.RRect? rrect, {
@@ -292,6 +306,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipPath(
     _i2.Path? path, {
@@ -305,6 +320,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i2.Rect getLocalClipBounds() => (super.noSuchMethod(
         Invocation.method(
@@ -319,6 +335,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
           ),
         ),
       ) as _i2.Rect);
+
   @override
   _i2.Rect getDestinationClipBounds() => (super.noSuchMethod(
         Invocation.method(
@@ -333,6 +350,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
           ),
         ),
       ) as _i2.Rect);
+
   @override
   void drawColor(
     _i2.Color? color,
@@ -348,6 +366,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawLine(
     _i2.Offset? p1,
@@ -365,6 +384,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPaint(_i2.Paint? paint) => super.noSuchMethod(
         Invocation.method(
@@ -373,6 +393,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRect(
     _i2.Rect? rect,
@@ -388,6 +409,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRRect(
     _i2.RRect? rrect,
@@ -403,6 +425,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawDRRect(
     _i2.RRect? outer,
@@ -420,6 +443,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawOval(
     _i2.Rect? rect,
@@ -435,6 +459,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawCircle(
     _i2.Offset? c,
@@ -452,6 +477,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawArc(
     _i2.Rect? rect,
@@ -473,6 +499,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPath(
     _i2.Path? path,
@@ -488,6 +515,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawImage(
     _i2.Image? image,
@@ -505,6 +533,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawImageRect(
     _i2.Image? image,
@@ -524,6 +553,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawImageNine(
     _i2.Image? image,
@@ -543,6 +573,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPicture(_i2.Picture? picture) => super.noSuchMethod(
         Invocation.method(
@@ -551,6 +582,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawParagraph(
     _i2.Paragraph? paragraph,
@@ -566,6 +598,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPoints(
     _i2.PointMode? pointMode,
@@ -583,6 +616,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRawPoints(
     _i2.PointMode? pointMode,
@@ -600,6 +634,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawVertices(
     _i2.Vertices? vertices,
@@ -617,6 +652,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawAtlas(
     _i2.Image? atlas,
@@ -642,6 +678,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRawAtlas(
     _i2.Image? atlas,
@@ -667,6 +704,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawShadow(
     _i2.Path? path,
@@ -704,6 +742,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           Invocation.getter(#estimatedBounds),
         ),
       ) as _i2.Rect);
+
   @override
   _i2.Canvas get canvas => (super.noSuchMethod(
         Invocation.getter(#canvas),
@@ -712,6 +751,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           Invocation.getter(#canvas),
         ),
       ) as _i2.Canvas);
+
   @override
   void paintChild(
     _i3.RenderObject? child,
@@ -727,6 +767,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void appendLayer(_i4.Layer? layer) => super.noSuchMethod(
         Invocation.method(
@@ -735,6 +776,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i2.VoidCallback addCompositionCallback(_i4.CompositionCallback? callback) =>
       (super.noSuchMethod(
@@ -744,6 +786,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValue: () {},
       ) as _i2.VoidCallback);
+
   @override
   void stopRecordingIfNeeded() => super.noSuchMethod(
         Invocation.method(
@@ -752,6 +795,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void setIsComplexHint() => super.noSuchMethod(
         Invocation.method(
@@ -760,6 +804,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void setWillChangeHint() => super.noSuchMethod(
         Invocation.method(
@@ -768,6 +813,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void addLayer(_i4.Layer? layer) => super.noSuchMethod(
         Invocation.method(
@@ -776,6 +822,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void pushLayer(
     _i4.ContainerLayer? childLayer,
@@ -795,6 +842,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i3.PaintingContext createChildContext(
     _i4.ContainerLayer? childLayer,
@@ -819,6 +867,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           ),
         ),
       ) as _i3.PaintingContext);
+
   @override
   _i4.ClipRectLayer? pushClipRect(
     bool? needsCompositing,
@@ -841,6 +890,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           #oldLayer: oldLayer,
         },
       )) as _i4.ClipRectLayer?);
+
   @override
   _i4.ClipRRectLayer? pushClipRRect(
     bool? needsCompositing,
@@ -865,6 +915,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           #oldLayer: oldLayer,
         },
       )) as _i4.ClipRRectLayer?);
+
   @override
   _i4.ClipPathLayer? pushClipPath(
     bool? needsCompositing,
@@ -889,6 +940,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           #oldLayer: oldLayer,
         },
       )) as _i4.ClipPathLayer?);
+
   @override
   _i4.ColorFilterLayer pushColorFilter(
     _i2.Offset? offset,
@@ -919,6 +971,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           ),
         ),
       ) as _i4.ColorFilterLayer);
+
   @override
   _i4.TransformLayer? pushTransform(
     bool? needsCompositing,
@@ -937,6 +990,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ],
         {#oldLayer: oldLayer},
       )) as _i4.TransformLayer?);
+
   @override
   _i4.OpacityLayer pushOpacity(
     _i2.Offset? offset,
@@ -967,6 +1021,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
           ),
         ),
       ) as _i4.OpacityLayer);
+
   @override
   void clipPathAndPaint(
     _i2.Path? path,
@@ -986,6 +1041,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipRRectAndPaint(
     _i2.RRect? rrect,
@@ -1005,6 +1061,7 @@ class MockPaintingContext extends _i1.Mock implements _i3.PaintingContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipRectAndPaint(
     _i2.Rect? rect,
@@ -1042,16 +1099,19 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
           Invocation.getter(#widget),
         ),
       ) as _i6.Widget);
+
   @override
   bool get mounted => (super.noSuchMethod(
         Invocation.getter(#mounted),
         returnValue: false,
       ) as bool);
+
   @override
   bool get debugDoingBuild => (super.noSuchMethod(
         Invocation.getter(#debugDoingBuild),
         returnValue: false,
       ) as bool);
+
   @override
   _i6.InheritedWidget dependOnInheritedElement(
     _i6.InheritedElement? ancestor, {
@@ -1072,6 +1132,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
           ),
         ),
       ) as _i6.InheritedWidget);
+
   @override
   void visitAncestorElements(_i6.ConditionalElementVisitor? visitor) =>
       super.noSuchMethod(
@@ -1081,6 +1142,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void visitChildElements(_i6.ElementVisitor? visitor) => super.noSuchMethod(
         Invocation.method(
@@ -1089,6 +1151,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void dispatchNotification(_i9.Notification? notification) =>
       super.noSuchMethod(
@@ -1098,6 +1161,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i5.DiagnosticsNode describeElement(
     String? name, {
@@ -1118,6 +1182,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
           ),
         ),
       ) as _i5.DiagnosticsNode);
+
   @override
   _i5.DiagnosticsNode describeWidget(
     String? name, {
@@ -1138,6 +1203,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
           ),
         ),
       ) as _i5.DiagnosticsNode);
+
   @override
   List<_i5.DiagnosticsNode> describeMissingAncestor(
           {required Type? expectedAncestorType}) =>
@@ -1149,6 +1215,7 @@ class MockBuildContext extends _i1.Mock implements _i6.BuildContext {
         ),
         returnValue: <_i5.DiagnosticsNode>[],
       ) as List<_i5.DiagnosticsNode>);
+
   @override
   _i5.DiagnosticsNode describeOwnershipChain(String? name) =>
       (super.noSuchMethod(
@@ -1183,6 +1250,7 @@ class MockRadarChartPainter extends _i1.Mock implements _i10.RadarChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void paint(
     _i6.BuildContext? context,
@@ -1200,6 +1268,7 @@ class MockRadarChartPainter extends _i1.Mock implements _i10.RadarChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   double getDefaultChartCenterValue() => (super.noSuchMethod(
         Invocation.method(
@@ -1208,6 +1277,7 @@ class MockRadarChartPainter extends _i1.Mock implements _i10.RadarChartPainter {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   double getChartCenterValue(_i13.RadarChartData? data) => (super.noSuchMethod(
         Invocation.method(
@@ -1216,6 +1286,7 @@ class MockRadarChartPainter extends _i1.Mock implements _i10.RadarChartPainter {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   double getScaledPoint(
     _i13.RadarEntry? point,
@@ -1233,6 +1304,7 @@ class MockRadarChartPainter extends _i1.Mock implements _i10.RadarChartPainter {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   double getFirstTickValue(_i13.RadarChartData? data) => (super.noSuchMethod(
         Invocation.method(
@@ -1241,6 +1313,7 @@ class MockRadarChartPainter extends _i1.Mock implements _i10.RadarChartPainter {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   double getSpaceBetweenTicks(_i13.RadarChartData? data) => (super.noSuchMethod(
         Invocation.method(
@@ -1249,6 +1322,7 @@ class MockRadarChartPainter extends _i1.Mock implements _i10.RadarChartPainter {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   void drawTicks(
     _i6.BuildContext? context,
@@ -1266,6 +1340,7 @@ class MockRadarChartPainter extends _i1.Mock implements _i10.RadarChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawGrids(
     _i11.CanvasWrapper? canvasWrapper,
@@ -1281,6 +1356,7 @@ class MockRadarChartPainter extends _i1.Mock implements _i10.RadarChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawTitles(
     _i6.BuildContext? context,
@@ -1298,6 +1374,7 @@ class MockRadarChartPainter extends _i1.Mock implements _i10.RadarChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawDataSets(
     _i11.CanvasWrapper? canvasWrapper,
@@ -1313,6 +1390,7 @@ class MockRadarChartPainter extends _i1.Mock implements _i10.RadarChartPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i13.RadarTouchedSpot? handleTouch(
     _i2.Offset? touchedPoint,
@@ -1327,6 +1405,7 @@ class MockRadarChartPainter extends _i1.Mock implements _i10.RadarChartPainter {
           holder,
         ],
       )) as _i13.RadarTouchedSpot?);
+
   @override
   double radarCenterY(_i2.Size? size) => (super.noSuchMethod(
         Invocation.method(
@@ -1335,6 +1414,7 @@ class MockRadarChartPainter extends _i1.Mock implements _i10.RadarChartPainter {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   double radarCenterX(_i2.Size? size) => (super.noSuchMethod(
         Invocation.method(
@@ -1343,6 +1423,7 @@ class MockRadarChartPainter extends _i1.Mock implements _i10.RadarChartPainter {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   double radarRadius(_i2.Size? size) => (super.noSuchMethod(
         Invocation.method(
@@ -1351,6 +1432,7 @@ class MockRadarChartPainter extends _i1.Mock implements _i10.RadarChartPainter {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   List<_i10.RadarDataSetsPosition> calculateDataSetsPosition(
     _i2.Size? viewSize,
