@@ -28,7 +28,11 @@ void main() {
       );
 
       final scatterPainter = ScatterChartPainter();
-      final holder = PaintHolder<ScatterChartData>(data, data, 1);
+      final holder = PaintHolder<ScatterChartData>(
+        data,
+        data,
+        TextScaler.noScaling,
+      );
 
       final mockUtils = MockUtils();
       Utils.changeInstance(mockUtils);
@@ -58,7 +62,7 @@ void main() {
         holder,
       );
 
-      verify(mockCanvasWrapper.drawCircle(any, any, any)).called(3);
+      verify(mockCanvasWrapper.drawDot(any, any, any)).called(3);
       Utils.changeInstance(utilsMainInstance);
     });
   });
@@ -67,23 +71,30 @@ void main() {
     test('test 1', () {
       const viewSize = Size(100, 100);
 
+      final dotPainter1 = FlDotCirclePainter(radius: 18);
+      final dotPainter3 = FlDotCirclePainter(radius: 4);
+      final dotPainter4 = FlDotCirclePainter(radius: 6);
+
+      final spot1 = ScatterSpot(1, 1, dotPainter: dotPainter1);
+      final spot2 = ScatterSpot(3, 9, show: false);
+      final spot3 = ScatterSpot(8, 2, dotPainter: dotPainter3);
+      final spot4 = ScatterSpot(7, 5, dotPainter: dotPainter4);
       final data = ScatterChartData(
         minY: 0,
         maxY: 10,
         minX: 0,
         maxX: 10,
-        scatterSpots: [
-          ScatterSpot(1, 1, radius: 18),
-          ScatterSpot(3, 9, show: false),
-          ScatterSpot(8, 2, radius: 4),
-          ScatterSpot(7, 5, radius: 6),
-        ],
+        scatterSpots: [spot1, spot2, spot3, spot4],
         titlesData: const FlTitlesData(show: false),
         clipData: const FlClipData.all(),
       );
 
       final scatterChartPainter = ScatterChartPainter();
-      final holder = PaintHolder<ScatterChartData>(data, data, 1);
+      final holder = PaintHolder<ScatterChartData>(
+        data,
+        data,
+        TextScaler.noScaling,
+      );
 
       final mockBuildContext = MockBuildContext();
       final mockCanvasWrapper = MockCanvasWrapper();
@@ -95,12 +106,27 @@ void main() {
         holder,
       );
 
-      verify(mockCanvasWrapper.drawCircle(const Offset(10, 90), 18, any))
-          .called(1);
-      verify(mockCanvasWrapper.drawCircle(const Offset(80, 80), 4, any))
-          .called(1);
-      verify(mockCanvasWrapper.drawCircle(const Offset(70, 50), 6, any))
-          .called(1);
+      verify(
+        mockCanvasWrapper.drawDot(
+          dotPainter1,
+          spot1,
+          const Offset(10, 90),
+        ),
+      ).called(1);
+      verify(
+        mockCanvasWrapper.drawDot(
+          dotPainter3,
+          spot3,
+          const Offset(80, 80),
+        ),
+      ).called(1);
+      verify(
+        mockCanvasWrapper.drawDot(
+          dotPainter4,
+          spot4,
+          const Offset(70, 50),
+        ),
+      ).called(1);
 
       verifyNever(mockCanvasWrapper.drawText(any, any));
       verify(mockCanvasWrapper.clipRect(any)).called(1);
@@ -125,7 +151,11 @@ void main() {
       );
 
       final scatterChartPainter = ScatterChartPainter();
-      final holder = PaintHolder<ScatterChartData>(data, data, 1);
+      final holder = PaintHolder<ScatterChartData>(
+        data,
+        data,
+        TextScaler.noScaling,
+      );
 
       final mockBuildContext = MockBuildContext();
       final mockCanvasWrapper = MockCanvasWrapper();
@@ -152,12 +182,12 @@ void main() {
         minX: 0,
         maxX: 10,
         scatterSpots: [
-          ScatterSpot(1, 1, radius: 18),
-          ScatterSpot(2, 2, radius: 8),
+          ScatterSpot(1, 1, dotPainter: FlDotCirclePainter(radius: 18)),
+          ScatterSpot(2, 2, dotPainter: FlDotCirclePainter(radius: 8)),
           ScatterSpot(3, 9, show: false),
-          ScatterSpot(8, 8, radius: 4),
-          ScatterSpot(7, 5, radius: 20),
-          ScatterSpot(4, 6, radius: 24),
+          ScatterSpot(8, 8, dotPainter: FlDotCirclePainter(radius: 4)),
+          ScatterSpot(7, 5, dotPainter: FlDotCirclePainter(radius: 20)),
+          ScatterSpot(4, 6, dotPainter: FlDotCirclePainter(radius: 24)),
         ],
         titlesData: const FlTitlesData(show: false),
         clipData: const FlClipData.all(),
@@ -175,7 +205,11 @@ void main() {
       );
 
       final scatterChartPainter = ScatterChartPainter();
-      final holder = PaintHolder<ScatterChartData>(data, data, 1);
+      final holder = PaintHolder<ScatterChartData>(
+        data,
+        data,
+        TextScaler.noScaling,
+      );
 
       final mockBuildContext = MockBuildContext();
       final mockCanvasWrapper = MockCanvasWrapper();
@@ -194,16 +228,41 @@ void main() {
         holder,
       );
 
-      verify(mockCanvasWrapper.drawCircle(const Offset(10, 90), 18, any))
-          .called(1);
-      verify(mockCanvasWrapper.drawCircle(const Offset(20, 80), 8, any))
-          .called(1);
-      verify(mockCanvasWrapper.drawCircle(const Offset(80, 20), 4, any))
-          .called(1);
-      verify(mockCanvasWrapper.drawCircle(const Offset(70, 50), 20, any))
-          .called(1);
-      verify(mockCanvasWrapper.drawCircle(const Offset(40, 40), 24, any))
-          .called(1);
+      verify(
+        mockCanvasWrapper.drawDot(
+          any,
+          data.scatterSpots[0],
+          const Offset(10, 90),
+        ),
+      ).called(1);
+      verify(
+        mockCanvasWrapper.drawDot(
+          any,
+          data.scatterSpots[1],
+          const Offset(20, 80),
+        ),
+      ).called(1);
+      verify(
+        mockCanvasWrapper.drawDot(
+          any,
+          data.scatterSpots[3],
+          const Offset(80, 20),
+        ),
+      ).called(1);
+      verify(
+        mockCanvasWrapper.drawDot(
+          any,
+          data.scatterSpots[4],
+          const Offset(70, 50),
+        ),
+      ).called(1);
+      verify(
+        mockCanvasWrapper.drawDot(
+          any,
+          data.scatterSpots[5],
+          const Offset(40, 40),
+        ),
+      ).called(1);
 
       verify(mockCanvasWrapper.drawText(any, any)).called(4);
 
@@ -221,17 +280,21 @@ void main() {
         minX: 0,
         maxX: 10,
         scatterSpots: [
-          ScatterSpot(1, 1, radius: 18),
+          ScatterSpot(1, 1, dotPainter: FlDotCirclePainter(radius: 18)),
           ScatterSpot(3, 9, show: false),
-          ScatterSpot(8, 2, radius: 4),
-          ScatterSpot(7, 5, radius: 6),
+          ScatterSpot(8, 2, dotPainter: FlDotCirclePainter(radius: 4)),
+          ScatterSpot(7, 5, dotPainter: FlDotCirclePainter(radius: 6)),
         ],
         showingTooltipIndicators: [0, 2, 3],
         titlesData: const FlTitlesData(show: false),
       );
 
       final scatterChartPainter = ScatterChartPainter();
-      final holder = PaintHolder<ScatterChartData>(data, data, 1);
+      final holder = PaintHolder<ScatterChartData>(
+        data,
+        data,
+        TextScaler.noScaling,
+      );
       final mockCanvasWrapper = MockCanvasWrapper();
       final mockBuildContext = MockBuildContext();
       final mockUtils = MockUtils();
@@ -267,10 +330,10 @@ void main() {
         minX: 0,
         maxX: 10,
         scatterSpots: [
-          ScatterSpot(1, 1, radius: 18),
+          ScatterSpot(1, 1, dotPainter: FlDotCirclePainter(radius: 18)),
           ScatterSpot(3, 9, show: false),
-          ScatterSpot(8, 2, radius: 4),
-          ScatterSpot(7, 5, radius: 6),
+          ScatterSpot(8, 2, dotPainter: FlDotCirclePainter(radius: 4)),
+          ScatterSpot(7, 5, dotPainter: FlDotCirclePainter(radius: 6)),
         ],
         showingTooltipIndicators: [0, 2, 3],
         scatterTouchData: ScatterTouchData(
@@ -282,7 +345,11 @@ void main() {
       );
 
       final scatterChartPainter = ScatterChartPainter();
-      final holder = PaintHolder<ScatterChartData>(data, data, 1);
+      final holder = PaintHolder<ScatterChartData>(
+        data,
+        data,
+        TextScaler.noScaling,
+      );
       final mockCanvasWrapper = MockCanvasWrapper();
       final mockBuildContext = MockBuildContext();
       final mockUtils = MockUtils();
@@ -350,7 +417,11 @@ void main() {
       );
 
       final scatterChartPainter = ScatterChartPainter();
-      final holder = PaintHolder<ScatterChartData>(data, data, 1);
+      final holder = PaintHolder<ScatterChartData>(
+        data,
+        data,
+        TextScaler.noScaling,
+      );
       final mockCanvasWrapper = MockCanvasWrapper();
       final mockBuildContext = MockBuildContext();
       final mockUtils = MockUtils();
@@ -451,7 +522,11 @@ void main() {
       );
 
       final scatterChartPainter = ScatterChartPainter();
-      final holder = PaintHolder<ScatterChartData>(data, data, 1);
+      final holder = PaintHolder<ScatterChartData>(
+        data,
+        data,
+        TextScaler.noScaling,
+      );
       final mockCanvasWrapper = MockCanvasWrapper();
       final mockBuildContext = MockBuildContext();
       final mockUtils = MockUtils();
@@ -554,7 +629,11 @@ void main() {
       );
 
       final scatterChartPainter = ScatterChartPainter();
-      final holder = PaintHolder<ScatterChartData>(data, data, 1);
+      final holder = PaintHolder<ScatterChartData>(
+        data,
+        data,
+        TextScaler.noScaling,
+      );
       final mockCanvasWrapper = MockCanvasWrapper();
       final mockBuildContext = MockBuildContext();
       final mockUtils = MockUtils();
@@ -621,7 +700,7 @@ void main() {
       final spots = [
         ScatterSpot(1, 1),
         ScatterSpot(2, 4),
-        ScatterSpot(5, 2, radius: 0.5),
+        ScatterSpot(5, 2, dotPainter: FlDotCirclePainter(radius: 0.5)),
         ScatterSpot(8, 7),
       ];
 
@@ -635,7 +714,11 @@ void main() {
       );
 
       final scatterChartPainter = ScatterChartPainter();
-      final holder = PaintHolder<ScatterChartData>(data, data, 1);
+      final holder = PaintHolder<ScatterChartData>(
+        data,
+        data,
+        TextScaler.noScaling,
+      );
       final touchedSpot = scatterChartPainter.handleTouch(
         const Offset(10, 90),
         viewSize,
@@ -664,7 +747,7 @@ void main() {
       );
       expect(touchedSpot4, null);
 
-      final radius = spots[2].radius;
+      final radius = spots[2].size.width / 2;
       final touchedSpot5 = scatterChartPainter.handleTouch(
         Offset(
           50 + (math.cos(math.pi / 4) * radius) - 0.01,
@@ -691,7 +774,7 @@ void main() {
       final spots = [
         ScatterSpot(1, 1),
         ScatterSpot(2, 4),
-        ScatterSpot(5, 2, radius: 0.5),
+        ScatterSpot(5, 2, dotPainter: FlDotCirclePainter(radius: 0.5)),
         ScatterSpot(8, 7),
       ];
 
@@ -734,7 +817,11 @@ void main() {
       );
 
       final scatterChartPainter = ScatterChartPainter();
-      final holder = PaintHolder<ScatterChartData>(data, data, 1);
+      final holder = PaintHolder<ScatterChartData>(
+        data,
+        data,
+        TextScaler.noScaling,
+      );
       final touchedSpot = scatterChartPainter.handleTouch(
         const Offset(10, 90),
         viewSize,
@@ -763,7 +850,7 @@ void main() {
       );
       expect(touchedSpot4, null);
 
-      final radius = spots[2].radius;
+      final radius = spots[2].size.width / 2;
       final touchedSpot5 = scatterChartPainter.handleTouch(
         Offset(
           50 + (math.cos(math.pi / 4) * radius) - 0.01,
