@@ -1,20 +1,23 @@
-// Mocks generated by Mockito 5.4.0 from annotations
+// Mocks generated by Mockito 5.4.3 from annotations
 // in fl_chart/test/utils/canvas_wrapper_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'dart:typed_data' as _i4;
+import 'dart:typed_data' as _i5;
 import 'dart:ui' as _i2;
 
-import 'package:fl_chart/fl_chart.dart' as _i5;
+import 'package:fl_chart/fl_chart.dart' as _i3;
 import 'package:fl_chart/src/utils/utils.dart' as _i6;
-import 'package:flutter/material.dart' as _i3;
+import 'package:flutter/material.dart' as _i4;
 import 'package:mockito/mockito.dart' as _i1;
+import 'package:mockito/src/dummies.dart' as _i7;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
 // ignore_for_file: avoid_setters_without_getters
 // ignore_for_file: comment_references
+// ignore_for_file: deprecated_member_use
+// ignore_for_file: deprecated_member_use_from_same_package
 // ignore_for_file: implementation_imports
 // ignore_for_file: invalid_use_of_visible_for_testing_member
 // ignore_for_file: prefer_const_constructors
@@ -32,8 +35,8 @@ class _FakeRect_0 extends _i1.SmartFake implements _i2.Rect {
         );
 }
 
-class _FakeSize_1 extends _i1.SmartFake implements _i2.Size {
-  _FakeSize_1(
+class _FakeColor_1 extends _i1.SmartFake implements _i2.Color {
+  _FakeColor_1(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -42,8 +45,8 @@ class _FakeSize_1 extends _i1.SmartFake implements _i2.Size {
         );
 }
 
-class _FakeOffset_2 extends _i1.SmartFake implements _i2.Offset {
-  _FakeOffset_2(
+class _FakeSize_2 extends _i1.SmartFake implements _i2.Size {
+  _FakeSize_2(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -52,8 +55,28 @@ class _FakeOffset_2 extends _i1.SmartFake implements _i2.Offset {
         );
 }
 
-class _FakeBorderSide_3 extends _i1.SmartFake implements _i3.BorderSide {
-  _FakeBorderSide_3(
+class _FakeFlDotPainter_3 extends _i1.SmartFake implements _i3.FlDotPainter {
+  _FakeFlDotPainter_3(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
+
+class _FakeOffset_4 extends _i1.SmartFake implements _i2.Offset {
+  _FakeOffset_4(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
+
+class _FakeBorderSide_5 extends _i1.SmartFake implements _i4.BorderSide {
+  _FakeBorderSide_5(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -62,12 +85,12 @@ class _FakeBorderSide_3 extends _i1.SmartFake implements _i3.BorderSide {
         );
 
   @override
-  String toString({_i3.DiagnosticLevel? minLevel = _i3.DiagnosticLevel.info}) =>
+  String toString({_i4.DiagnosticLevel? minLevel = _i4.DiagnosticLevel.info}) =>
       super.toString();
 }
 
-class _FakeTextStyle_4 extends _i1.SmartFake implements _i3.TextStyle {
-  _FakeTextStyle_4(
+class _FakeTextStyle_6 extends _i1.SmartFake implements _i4.TextStyle {
+  _FakeTextStyle_6(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -76,7 +99,7 @@ class _FakeTextStyle_4 extends _i1.SmartFake implements _i3.TextStyle {
         );
 
   @override
-  String toString({_i3.DiagnosticLevel? minLevel = _i3.DiagnosticLevel.info}) =>
+  String toString({_i4.DiagnosticLevel? minLevel = _i4.DiagnosticLevel.info}) =>
       super.toString();
 }
 
@@ -96,6 +119,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void saveLayer(
     _i2.Rect? bounds,
@@ -111,6 +135,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void restore() => super.noSuchMethod(
         Invocation.method(
@@ -119,6 +144,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void restoreToCount(int? count) => super.noSuchMethod(
         Invocation.method(
@@ -127,6 +153,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   int getSaveCount() => (super.noSuchMethod(
         Invocation.method(
@@ -135,6 +162,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValue: 0,
       ) as int);
+
   @override
   void translate(
     double? dx,
@@ -150,6 +178,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void scale(
     double? sx, [
@@ -165,6 +194,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void rotate(double? radians) => super.noSuchMethod(
         Invocation.method(
@@ -173,6 +203,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void skew(
     double? sx,
@@ -188,22 +219,25 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
-  void transform(_i4.Float64List? matrix4) => super.noSuchMethod(
+  void transform(_i5.Float64List? matrix4) => super.noSuchMethod(
         Invocation.method(
           #transform,
           [matrix4],
         ),
         returnValueForMissingStub: null,
       );
+
   @override
-  _i4.Float64List getTransform() => (super.noSuchMethod(
+  _i5.Float64List getTransform() => (super.noSuchMethod(
         Invocation.method(
           #getTransform,
           [],
         ),
-        returnValue: _i4.Float64List(0),
-      ) as _i4.Float64List);
+        returnValue: _i5.Float64List(0),
+      ) as _i5.Float64List);
+
   @override
   void clipRect(
     _i2.Rect? rect, {
@@ -221,6 +255,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipRRect(
     _i2.RRect? rrect, {
@@ -234,6 +269,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clipPath(
     _i2.Path? path, {
@@ -247,6 +283,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i2.Rect getLocalClipBounds() => (super.noSuchMethod(
         Invocation.method(
@@ -261,6 +298,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
           ),
         ),
       ) as _i2.Rect);
+
   @override
   _i2.Rect getDestinationClipBounds() => (super.noSuchMethod(
         Invocation.method(
@@ -275,6 +313,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
           ),
         ),
       ) as _i2.Rect);
+
   @override
   void drawColor(
     _i2.Color? color,
@@ -290,6 +329,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawLine(
     _i2.Offset? p1,
@@ -307,6 +347,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPaint(_i2.Paint? paint) => super.noSuchMethod(
         Invocation.method(
@@ -315,6 +356,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRect(
     _i2.Rect? rect,
@@ -330,6 +372,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRRect(
     _i2.RRect? rrect,
@@ -345,6 +388,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawDRRect(
     _i2.RRect? outer,
@@ -362,6 +406,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawOval(
     _i2.Rect? rect,
@@ -377,6 +422,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawCircle(
     _i2.Offset? c,
@@ -394,6 +440,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawArc(
     _i2.Rect? rect,
@@ -415,6 +462,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPath(
     _i2.Path? path,
@@ -430,6 +478,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawImage(
     _i2.Image? image,
@@ -447,6 +496,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawImageRect(
     _i2.Image? image,
@@ -466,6 +516,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawImageNine(
     _i2.Image? image,
@@ -485,6 +536,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPicture(_i2.Picture? picture) => super.noSuchMethod(
         Invocation.method(
@@ -493,6 +545,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawParagraph(
     _i2.Paragraph? paragraph,
@@ -508,6 +561,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawPoints(
     _i2.PointMode? pointMode,
@@ -525,10 +579,11 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRawPoints(
     _i2.PointMode? pointMode,
-    _i4.Float32List? points,
+    _i5.Float32List? points,
     _i2.Paint? paint,
   ) =>
       super.noSuchMethod(
@@ -542,6 +597,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawVertices(
     _i2.Vertices? vertices,
@@ -559,6 +615,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawAtlas(
     _i2.Image? atlas,
@@ -584,12 +641,13 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawRawAtlas(
     _i2.Image? atlas,
-    _i4.Float32List? rstTransforms,
-    _i4.Float32List? rects,
-    _i4.Int32List? colors,
+    _i5.Float32List? rstTransforms,
+    _i5.Float32List? rects,
+    _i5.Int32List? colors,
     _i2.BlendMode? blendMode,
     _i2.Rect? cullRect,
     _i2.Paint? paint,
@@ -609,6 +667,7 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void drawShadow(
     _i2.Path? path,
@@ -633,20 +692,30 @@ class MockCanvas extends _i1.Mock implements _i2.Canvas {
 /// A class which mocks [FlDotPainter].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockFlDotPainter extends _i1.Mock implements _i5.FlDotPainter {
+class MockFlDotPainter extends _i1.Mock implements _i3.FlDotPainter {
   MockFlDotPainter() {
     _i1.throwOnMissingStub(this);
   }
+
+  @override
+  _i2.Color get mainColor => (super.noSuchMethod(
+        Invocation.getter(#mainColor),
+        returnValue: _FakeColor_1(
+          this,
+          Invocation.getter(#mainColor),
+        ),
+      ) as _i2.Color);
 
   @override
   List<Object?> get props => (super.noSuchMethod(
         Invocation.getter(#props),
         returnValue: <Object?>[],
       ) as List<Object?>);
+
   @override
   void draw(
     _i2.Canvas? canvas,
-    _i5.FlSpot? spot,
+    _i3.FlSpot? spot,
     _i2.Offset? offsetInCanvas,
   ) =>
       super.noSuchMethod(
@@ -660,13 +729,14 @@ class MockFlDotPainter extends _i1.Mock implements _i5.FlDotPainter {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
-  _i2.Size getSize(_i5.FlSpot? spot) => (super.noSuchMethod(
+  _i2.Size getSize(_i3.FlSpot? spot) => (super.noSuchMethod(
         Invocation.method(
           #getSize,
           [spot],
         ),
-        returnValue: _FakeSize_1(
+        returnValue: _FakeSize_2(
           this,
           Invocation.method(
             #getSize,
@@ -674,6 +744,54 @@ class MockFlDotPainter extends _i1.Mock implements _i5.FlDotPainter {
           ),
         ),
       ) as _i2.Size);
+
+  @override
+  _i3.FlDotPainter lerp(
+    _i3.FlDotPainter? a,
+    _i3.FlDotPainter? b,
+    double? t,
+  ) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #lerp,
+          [
+            a,
+            b,
+            t,
+          ],
+        ),
+        returnValue: _FakeFlDotPainter_3(
+          this,
+          Invocation.method(
+            #lerp,
+            [
+              a,
+              b,
+              t,
+            ],
+          ),
+        ),
+      ) as _i3.FlDotPainter);
+
+  @override
+  bool hitTest(
+    _i3.FlSpot? spot,
+    _i2.Offset? touched,
+    _i2.Offset? center,
+    double? extraThreshold,
+  ) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #hitTest,
+          [
+            spot,
+            touched,
+            center,
+            extraThreshold,
+          ],
+        ),
+        returnValue: false,
+      ) as bool);
 }
 
 /// A class which mocks [Utils].
@@ -692,6 +810,7 @@ class MockUtils extends _i1.Mock implements _i6.Utils {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   double degrees(double? radians) => (super.noSuchMethod(
         Invocation.method(
@@ -700,13 +819,14 @@ class MockUtils extends _i1.Mock implements _i6.Utils {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   _i2.Size getDefaultSize(_i2.Size? screenSize) => (super.noSuchMethod(
         Invocation.method(
           #getDefaultSize,
           [screenSize],
         ),
-        returnValue: _FakeSize_1(
+        returnValue: _FakeSize_2(
           this,
           Invocation.method(
             #getDefaultSize,
@@ -714,6 +834,7 @@ class MockUtils extends _i1.Mock implements _i6.Utils {
           ),
         ),
       ) as _i2.Size);
+
   @override
   double translateRotatedPosition(
     double? size,
@@ -729,6 +850,7 @@ class MockUtils extends _i1.Mock implements _i6.Utils {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   _i2.Offset calculateRotationOffset(
     _i2.Size? size,
@@ -742,7 +864,7 @@ class MockUtils extends _i1.Mock implements _i6.Utils {
             degree,
           ],
         ),
-        returnValue: _FakeOffset_2(
+        returnValue: _FakeOffset_4(
           this,
           Invocation.method(
             #calculateRotationOffset,
@@ -753,9 +875,10 @@ class MockUtils extends _i1.Mock implements _i6.Utils {
           ),
         ),
       ) as _i2.Offset);
+
   @override
-  _i3.BorderRadius? normalizeBorderRadius(
-    _i3.BorderRadius? borderRadius,
+  _i4.BorderRadius? normalizeBorderRadius(
+    _i4.BorderRadius? borderRadius,
     double? width,
   ) =>
       (super.noSuchMethod(Invocation.method(
@@ -764,10 +887,11 @@ class MockUtils extends _i1.Mock implements _i6.Utils {
           borderRadius,
           width,
         ],
-      )) as _i3.BorderRadius?);
+      )) as _i4.BorderRadius?);
+
   @override
-  _i3.BorderSide normalizeBorderSide(
-    _i3.BorderSide? borderSide,
+  _i4.BorderSide normalizeBorderSide(
+    _i4.BorderSide? borderSide,
     double? width,
   ) =>
       (super.noSuchMethod(
@@ -778,7 +902,7 @@ class MockUtils extends _i1.Mock implements _i6.Utils {
             width,
           ],
         ),
-        returnValue: _FakeBorderSide_3(
+        returnValue: _FakeBorderSide_5(
           this,
           Invocation.method(
             #normalizeBorderSide,
@@ -788,7 +912,8 @@ class MockUtils extends _i1.Mock implements _i6.Utils {
             ],
           ),
         ),
-      ) as _i3.BorderSide);
+      ) as _i4.BorderSide);
+
   @override
   double getEfficientInterval(
     double? axisViewSize,
@@ -806,6 +931,7 @@ class MockUtils extends _i1.Mock implements _i6.Utils {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   double roundInterval(double? input) => (super.noSuchMethod(
         Invocation.method(
@@ -814,18 +940,48 @@ class MockUtils extends _i1.Mock implements _i6.Utils {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
-  String formatNumber(double? number) => (super.noSuchMethod(
+  int getFractionDigits(double? value) => (super.noSuchMethod(
+        Invocation.method(
+          #getFractionDigits,
+          [value],
+        ),
+        returnValue: 0,
+      ) as int);
+
+  @override
+  String formatNumber(
+    double? axisMin,
+    double? axisMax,
+    double? axisValue,
+  ) =>
+      (super.noSuchMethod(
         Invocation.method(
           #formatNumber,
-          [number],
+          [
+            axisMin,
+            axisMax,
+            axisValue,
+          ],
         ),
-        returnValue: '',
+        returnValue: _i7.dummyValue<String>(
+          this,
+          Invocation.method(
+            #formatNumber,
+            [
+              axisMin,
+              axisMax,
+              axisValue,
+            ],
+          ),
+        ),
       ) as String);
+
   @override
-  _i3.TextStyle getThemeAwareTextStyle(
-    _i3.BuildContext? context,
-    _i3.TextStyle? providedStyle,
+  _i4.TextStyle getThemeAwareTextStyle(
+    _i4.BuildContext? context,
+    _i4.TextStyle? providedStyle,
   ) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -835,7 +991,7 @@ class MockUtils extends _i1.Mock implements _i6.Utils {
             providedStyle,
           ],
         ),
-        returnValue: _FakeTextStyle_4(
+        returnValue: _FakeTextStyle_6(
           this,
           Invocation.method(
             #getThemeAwareTextStyle,
@@ -845,7 +1001,8 @@ class MockUtils extends _i1.Mock implements _i6.Utils {
             ],
           ),
         ),
-      ) as _i3.TextStyle);
+      ) as _i4.TextStyle);
+
   @override
   double getBestInitialIntervalValue(
     double? min,
@@ -865,6 +1022,7 @@ class MockUtils extends _i1.Mock implements _i6.Utils {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   double convertRadiusToSigma(double? radius) => (super.noSuchMethod(
         Invocation.method(
