@@ -1,4 +1,4 @@
-// Mocks generated by Mockito 5.4.0 from annotations
+// Mocks generated by Mockito 5.4.3 from annotations
 // in fl_chart/test/utils/utils_test.dart.
 // Do not manually edit this file.
 
@@ -9,11 +9,14 @@ import 'package:fl_chart/src/utils/utils.dart' as _i5;
 import 'package:flutter/foundation.dart' as _i4;
 import 'package:flutter/material.dart' as _i3;
 import 'package:mockito/mockito.dart' as _i1;
+import 'package:mockito/src/dummies.dart' as _i6;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
 // ignore_for_file: avoid_setters_without_getters
 // ignore_for_file: comment_references
+// ignore_for_file: deprecated_member_use
+// ignore_for_file: deprecated_member_use_from_same_package
 // ignore_for_file: implementation_imports
 // ignore_for_file: invalid_use_of_visible_for_testing_member
 // ignore_for_file: prefer_const_constructors
@@ -132,6 +135,7 @@ class MockUtils extends _i1.Mock implements _i5.Utils {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   double degrees(double? radians) => (super.noSuchMethod(
         Invocation.method(
@@ -140,6 +144,7 @@ class MockUtils extends _i1.Mock implements _i5.Utils {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   _i2.Size getDefaultSize(_i2.Size? screenSize) => (super.noSuchMethod(
         Invocation.method(
@@ -154,6 +159,7 @@ class MockUtils extends _i1.Mock implements _i5.Utils {
           ),
         ),
       ) as _i2.Size);
+
   @override
   double translateRotatedPosition(
     double? size,
@@ -169,6 +175,7 @@ class MockUtils extends _i1.Mock implements _i5.Utils {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   _i2.Offset calculateRotationOffset(
     _i2.Size? size,
@@ -193,6 +200,7 @@ class MockUtils extends _i1.Mock implements _i5.Utils {
           ),
         ),
       ) as _i2.Offset);
+
   @override
   _i3.BorderRadius? normalizeBorderRadius(
     _i3.BorderRadius? borderRadius,
@@ -205,6 +213,7 @@ class MockUtils extends _i1.Mock implements _i5.Utils {
           width,
         ],
       )) as _i3.BorderRadius?);
+
   @override
   _i3.BorderSide normalizeBorderSide(
     _i3.BorderSide? borderSide,
@@ -229,6 +238,7 @@ class MockUtils extends _i1.Mock implements _i5.Utils {
           ),
         ),
       ) as _i3.BorderSide);
+
   @override
   double getEfficientInterval(
     double? axisViewSize,
@@ -246,6 +256,7 @@ class MockUtils extends _i1.Mock implements _i5.Utils {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   double roundInterval(double? input) => (super.noSuchMethod(
         Invocation.method(
@@ -254,14 +265,44 @@ class MockUtils extends _i1.Mock implements _i5.Utils {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
-  String formatNumber(double? number) => (super.noSuchMethod(
+  int getFractionDigits(double? value) => (super.noSuchMethod(
+        Invocation.method(
+          #getFractionDigits,
+          [value],
+        ),
+        returnValue: 0,
+      ) as int);
+
+  @override
+  String formatNumber(
+    double? axisMin,
+    double? axisMax,
+    double? axisValue,
+  ) =>
+      (super.noSuchMethod(
         Invocation.method(
           #formatNumber,
-          [number],
+          [
+            axisMin,
+            axisMax,
+            axisValue,
+          ],
         ),
-        returnValue: '',
+        returnValue: _i6.dummyValue<String>(
+          this,
+          Invocation.method(
+            #formatNumber,
+            [
+              axisMin,
+              axisMax,
+              axisValue,
+            ],
+          ),
+        ),
       ) as String);
+
   @override
   _i3.TextStyle getThemeAwareTextStyle(
     _i3.BuildContext? context,
@@ -286,6 +327,7 @@ class MockUtils extends _i1.Mock implements _i5.Utils {
           ),
         ),
       ) as _i3.TextStyle);
+
   @override
   double getBestInitialIntervalValue(
     double? min,
@@ -305,6 +347,7 @@ class MockUtils extends _i1.Mock implements _i5.Utils {
         ),
         returnValue: 0.0,
       ) as double);
+
   @override
   double convertRadiusToSigma(double? radius) => (super.noSuchMethod(
         Invocation.method(
@@ -331,16 +374,19 @@ class MockBuildContext extends _i1.Mock implements _i3.BuildContext {
           Invocation.getter(#widget),
         ),
       ) as _i3.Widget);
+
   @override
   bool get mounted => (super.noSuchMethod(
         Invocation.getter(#mounted),
         returnValue: false,
       ) as bool);
+
   @override
   bool get debugDoingBuild => (super.noSuchMethod(
         Invocation.getter(#debugDoingBuild),
         returnValue: false,
       ) as bool);
+
   @override
   _i3.InheritedWidget dependOnInheritedElement(
     _i3.InheritedElement? ancestor, {
@@ -361,6 +407,7 @@ class MockBuildContext extends _i1.Mock implements _i3.BuildContext {
           ),
         ),
       ) as _i3.InheritedWidget);
+
   @override
   void visitAncestorElements(_i3.ConditionalElementVisitor? visitor) =>
       super.noSuchMethod(
@@ -370,6 +417,7 @@ class MockBuildContext extends _i1.Mock implements _i3.BuildContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void visitChildElements(_i3.ElementVisitor? visitor) => super.noSuchMethod(
         Invocation.method(
@@ -378,6 +426,7 @@ class MockBuildContext extends _i1.Mock implements _i3.BuildContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void dispatchNotification(_i3.Notification? notification) =>
       super.noSuchMethod(
@@ -387,6 +436,7 @@ class MockBuildContext extends _i1.Mock implements _i3.BuildContext {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i3.DiagnosticsNode describeElement(
     String? name, {
@@ -407,6 +457,7 @@ class MockBuildContext extends _i1.Mock implements _i3.BuildContext {
           ),
         ),
       ) as _i3.DiagnosticsNode);
+
   @override
   _i3.DiagnosticsNode describeWidget(
     String? name, {
@@ -427,6 +478,7 @@ class MockBuildContext extends _i1.Mock implements _i3.BuildContext {
           ),
         ),
       ) as _i3.DiagnosticsNode);
+
   @override
   List<_i3.DiagnosticsNode> describeMissingAncestor(
           {required Type? expectedAncestorType}) =>
@@ -438,6 +490,7 @@ class MockBuildContext extends _i1.Mock implements _i3.BuildContext {
         ),
         returnValue: <_i3.DiagnosticsNode>[],
       ) as List<_i3.DiagnosticsNode>);
+
   @override
   _i3.DiagnosticsNode describeOwnershipChain(String? name) =>
       (super.noSuchMethod(
